Organize network diagnostics tests and fix continuation leak
