+++ WiFiLens/Tests/WiFiLensTests/EditionCompositionTests.swift
@@ -284,8 +284,8 @@ struct EditionCompositionTests {
         let replyProbe = TerminationBoolProbe()
         var steps: [String] = []
         let coordinator = ApplicationTerminationCoordinator(
-            waitForDeadline: { _ in
-                await withCheckedContinuation { (_: CheckedContinuation<Void, Never>) in }
+            waitForDeadline: { duration in
+                try await Task.sleep(for: duration)
             },
             reply: { replyProbe.record($0) }
         )
