+++ WiFiLens/Tests/WiFiLensTests/NetworkDiagnostics/NetworkDiagnosticsTests.swift
@@ -4,3480 +4,10 @@ import Network
 import Testing
 @testable import WiFi_Lens
 
-@Suite("Network diagnostics models and runner")
+@Suite("Network diagnostics")
 struct NetworkDiagnosticsTests {
-    @Test("all app configurations use the same local-network privacy copy")
-    func localNetworkUsageDescriptionIsUnifiedAcrossAppConfigurations() throws {
-        let descriptions = try appLocalNetworkUsageDescriptions()
-        let expectedDescription = "WiFi Lens uses local networking when you enable its MCP server and when Network Self-Check tests reachability of your configured network proxy. These checks do not collect or transmit your Wi-Fi scan data."
 
-        #expect(descriptions.count == 4)
-        #expect(Set(descriptions.map(\.baseConfiguration)) == ["OSS.xcconfig", "PRO.xcconfig"])
-        #expect(descriptions.filter { $0.baseConfiguration == "OSS.xcconfig" }.count == 2)
-        #expect(descriptions.filter { $0.baseConfiguration == "PRO.xcconfig" }.count == 2)
-        #expect(Set(descriptions.map(\.value)) == [expectedDescription])
-        #expect(descriptions.allSatisfy { $0.value.contains("MCP server") })
-        #expect(descriptions.allSatisfy { $0.value.contains("Network Self-Check") })
-    }
-
-    @Test("privacy copy selection excludes decoy non-app configurations")
-    func localNetworkUsageDescriptionSelectionExcludesDecoyConfiguration() throws {
-        let descriptions = try appLocalNetworkUsageDescriptions(from: privacyCopyProjectFixture)
-
-        #expect(descriptions.count == 4)
-        #expect(descriptions.allSatisfy { $0.value == "expected app copy" })
-    }
-
-    @Test("diagnostic identifiers keep path and internet distinct in execution order")
-    func diagnosticIdentifierOrder() {
-        #expect(NetworkDiagnosticCheckID.allCases == [.path, .gatewayReachability, .dns, .internet, .ipv6, .proxy])
-        #expect(NetworkDiagnosticCheckID.path != .internet)
-        #expect(Set(NetworkDiagnosticCheckID.allCases).count == 6)
-    }
-
-    @Test("route selection follows the kernel-selected interface")
-    func routeSelectionUsesKernelInterface() {
-        let output = """
-           route to: default
-        destination: default
-               mask: default
-            gateway: 192.0.2.1
-          interface: en7
-              flags: <UP,GATEWAY,DONE,STATIC,PRCLONING,GLOBAL>
-        """
-
-        let result = DiagnosticRouteParser.parse(
-            output: output,
-            interfaceIndices: ["en0": 4, "en7": 12]
-        )
-
-        #expect(result == .selected(.init(
-            interfaceName: "en7",
-            interfaceIndex: 12,
-            address: "192.0.2.1"
-        )))
-    }
-
-    @Test("missing selected interface does not fall back to another adapter")
-    func missingInterfaceCannotSelectAnotherAdapter() {
-        let output = "destination: default\ngateway: 192.0.2.1\ninterface: en7\nflags: <UP,GATEWAY>"
-
-        #expect(DiagnosticRouteParser.parse(
-            output: output,
-            interfaceIndices: ["en0": 4]
-        ) == .unavailable)
-    }
-
-    @Test("route parsing is independent of interface enumeration order")
-    func routeSelectionIgnoresInterfaceEnumerationOrder() {
-        let output = "destination: default\ngateway: 192.0.2.1\ninterface: en7\nflags: <UP,GATEWAY>"
-        let firstOrder = DiagnosticRouteParser.parse(
-            output: output,
-            interfaceIndices: ["en0": 4, "en7": 12]
-        )
-        let reversedOrder = DiagnosticRouteParser.parse(
-            output: output,
-            interfaceIndices: ["en7": 12, "en0": 4]
-        )
-
-        #expect(firstOrder == reversedOrder)
-        #expect(firstOrder == .selected(.init(
-            interfaceName: "en7",
-            interfaceIndex: 12,
-            address: "192.0.2.1"
-        )))
-    }
-
-    @Test("same gateway address remains tied to the selected interface")
-    func sameGatewayAddressDoesNotCollapseInterfaceIdentity() {
-        let output = "destination: default\ngateway: 192.0.2.1\ninterface: en7\nflags: <UP,GATEWAY>"
-
-        #expect(DiagnosticRouteParser.parse(
-            output: output,
-            interfaceIndices: ["en0": 4, "en7": 12]
-        ) == .selected(.init(
-            interfaceName: "en7",
-            interfaceIndex: 12,
-            address: "192.0.2.1"
-        )))
-    }
-
-    @Test("duplicate route fields cannot produce a selected target")
-    func duplicateRouteFieldsAreRejected() {
-        let output = """
-        destination: default
-        gateway: 192.0.2.1
-        gateway: 192.0.2.254
-        interface: en7
-        flags: <UP,GATEWAY>
-        """
-
-        #expect(DiagnosticRouteParser.parse(
-            output: output,
-            interfaceIndices: ["en7": 12]
-        ) == .ambiguous)
-    }
-
-    @Test("invalid gateway addresses cannot be selected")
-    func invalidGatewayAddressesAreRejected() {
-        let addresses = ["0.0.0.0", "224.0.0.1", "255.255.255.255", "not-an-ip"]
-
-        for address in addresses {
-            let output = "destination: default\ngateway: \(address)\ninterface: en7\nflags: <UP,GATEWAY>"
-            let result = DiagnosticRouteParser.parse(
-                output: output,
-                interfaceIndices: ["en7": 12]
-            )
-            #expect(result == .unavailable)
-        }
-    }
-
-    @Test("link-layer gateways and tunnel interfaces are unsupported")
-    func unsupportedRouteTargetsAreRejected() {
-        let linkLayer = "destination: default\ngateway: link#4\ninterface: en7\nflags: <UP,GATEWAY>"
-        let tunnel = "destination: default\ngateway: 192.0.2.1\ninterface: utun3\nflags: <UP,GATEWAY>"
-
-        #expect(DiagnosticRouteParser.parse(
-            output: linkLayer,
-            interfaceIndices: ["en7": 12]
-        ) == .unsupported)
-        #expect(DiagnosticRouteParser.parse(
-            output: tunnel,
-            interfaceIndices: ["utun3": 20]
-        ) == .unsupported)
-    }
-
-    @Test("missing default route is unavailable")
-    func missingDefaultRouteIsUnavailable() {
-        let output = "gateway: 192.0.2.1\ninterface: en7\nflags: <UP,GATEWAY>"
-
-        #expect(DiagnosticRouteParser.parse(
-            output: output,
-            interfaceIndices: ["en7": 12]
-        ) == .unavailable)
-    }
-
-    @Test("route parser requires both UP and GATEWAY flags")
-    func routeParserRequiresUsableDefaultFlags() {
-        let records = [
-            "destination: default\ngateway: 192.0.2.1\ninterface: en7\nflags: <GATEWAY>",
-            "destination: default\ngateway: 192.0.2.1\ninterface: en7\nflags: <UP>",
-            "destination: default\ngateway: 192.0.2.1\ninterface: en7\nflags: <>"
-        ]
-
-        for output in records {
-            #expect(DiagnosticRouteParser.parse(
-                output: output,
-                interfaceIndices: ["en7": 12]
-            ) == .unavailable)
-        }
-    }
-
-    @Test("route command timeout terminates the child without waiting for EOF")
-    func routeCommandTimeoutDoesNotWaitForChild() async {
-        let execution = DiagnosticRouteProcessExecution(
-            executablePath: "/bin/sleep",
-            arguments: ["5"],
-            environment: [:],
-            outputLimit: 1024
-        )
-        let startedAt = ContinuousClock.now
-        let result = await execution.run(timeout: .milliseconds(50))
-        let elapsed = startedAt.duration(to: ContinuousClock.now)
-
-        #expect(result.timedOut)
-        #expect(elapsed < .seconds(2))
-    }
-
-    @Test("diagnostic gateway ping binds the selected interface")
-    func diagnosticGatewayPingBindsSelectedInterface() async {
-        let runner = RecordingGatewayPingProcessRunner(latency: 2.5)
-        let pinger = GatewayPinger(processRunner: runner)
-        let target = DiagnosticGatewayTarget(
-            interfaceName: "en7",
-            interfaceIndex: 12,
-            address: "192.0.2.1"
-        )
-
-        _ = await pinger.ping(target: target)
-
-        #expect(await runner.executablePath == "/sbin/ping")
-        #expect(await runner.arguments == [
-            "-b", "en7", "-c", "1", "-W", "1000", "192.0.2.1"
-        ])
-    }
-
-    @Test("overlapping ping runs preserve the latest cancellation owner")
-    func overlappingPingRunsPreserveCancellationOwnership() async {
-        let runner = SystemGatewayPingProcessRunner()
-        let first = Task {
-            await runner.run(executablePath: "/bin/sleep", arguments: ["0.1"])
-        }
-
-        try? await Task.sleep(for: .milliseconds(50))
-        let secondStartedAt = ContinuousClock.now
-        let second = Task {
-            await runner.run(executablePath: "/bin/sleep", arguments: ["10"])
-        }
-
-        try? await Task.sleep(for: .milliseconds(400))
-        await runner.cancel()
-
-        #expect(await first.value == nil)
-        #expect(await second.value == nil)
-        #expect(secondStartedAt.duration(to: ContinuousClock.now) < .seconds(1))
-    }
-
-    @Test("contextual path and gateway checks use one selected interface")
-    func contextualChecksShareSelectedRouteTarget() async {
-        let context = makeDiagnosticContext(
-            pathState: .satisfied,
-            route: .selected(.init(
-                interfaceName: "en7",
-                interfaceIndex: 12,
-                address: "192.0.2.1"
-            )),
-            interfaces: [
-                makeNetworkInterface(name: "en0", router: "192.0.2.1"),
-                makeNetworkInterface(name: "en7", router: "192.0.2.1"),
-            ]
-        )
-        let gateway = RecordingDiagnosticGatewayMeasurer()
-
-        let pathResult = await NetworkConnectivityCheck(context: context).run()
-        let gatewayResult = await GatewayReachabilityCheck(
-            context: context,
-            gatewayMeasuring: gateway,
-            routeSource: nil
-        ).run()
-
-        #expect(pathResult.evidence.contains(.init(code: "path.interface", value: "en7")))
-        #expect(pathResult.evidence.contains(.init(code: "path.gateway", value: "192.0.2.1")))
-        #expect(gatewayResult.evidence.contains(.init(code: "gateway.interface", value: "en7")))
-        #expect(await gateway.targets == [
-            .init(interfaceName: "en7", interfaceIndex: 12, address: "192.0.2.1")
-        ])
-    }
-
-    @Test("context capture accepts a stable route")
-    func diagnosticContextCaptureAcceptsStableRoute() async {
-        let route = DiagnosticRouteSelection.selected(.init(
-            interfaceName: "en7",
-            interfaceIndex: 12,
-            address: "192.0.2.1"
-        ))
-        let routeSource = SequencedDiagnosticRouteSource(values: [route, route])
-        let source = SystemDiagnosticNetworkContextSource(
-            routeSource: routeSource,
-            pathSource: StubPathSource(.satisfied),
-            interfaceSource: StubNetworkInterfaceSnapshotSource(interfaces: [])
-        )
-
-        let context = await source.capture(runID: UUID(), timeout: .seconds(1))
-
-        #expect(context?.route == route)
-        #expect(context?.pathState == .satisfied)
-        #expect(await routeSource.invocationCount == 2)
-    }
-
-    @Test("context capture does not publish a conflicting route")
-    func diagnosticContextCaptureRejectsChangingRoute() async {
-        let first = DiagnosticRouteSelection.selected(.init(
-            interfaceName: "en0",
-            interfaceIndex: 4,
-            address: "192.0.2.1"
-        ))
-        let second = DiagnosticRouteSelection.selected(.init(
-            interfaceName: "en7",
-            interfaceIndex: 12,
-            address: "192.0.2.1"
-        ))
-        let routeSource = SequencedDiagnosticRouteSource(values: [first, second, first, second])
-        let source = SystemDiagnosticNetworkContextSource(
-            routeSource: routeSource,
-            pathSource: StubPathSource(.satisfied),
-            interfaceSource: StubNetworkInterfaceSnapshotSource(interfaces: [])
-        )
-
-        let context = await source.capture(runID: UUID(), timeout: .seconds(1))
-
-        #expect(context?.route == .ambiguous)
-        #expect(await routeSource.invocationCount == 4)
-    }
-
-    @Test("route change during gateway ping suppresses the old result")
-    func gatewayRouteChangeSuppressesStaleResult() async {
-        let target = DiagnosticGatewayTarget(
-            interfaceName: "en0",
-            interfaceIndex: 4,
-            address: "192.0.2.1"
-        )
-        let context = makeDiagnosticContext(
-            pathState: .satisfied,
-            route: .selected(target)
-        )
-        let gateway = RecordingDiagnosticGatewayMeasurer()
-        let routeSource = SequencedDiagnosticRouteSource(values: [.unavailable])
-
-        let result = await GatewayReachabilityCheck(
-            context: context,
-            gatewayMeasuring: gateway,
-            routeSource: routeSource
-        ).run()
-
-        #expect(result.status == .indeterminate)
-        #expect(result.evidence.contains(.init(code: "gateway.route-changed", value: nil)))
-    }
-
-    @Test("missing diagnostic route never starts a gateway ping")
-    func missingDiagnosticRouteDoesNotPing() async {
-        let context = makeDiagnosticContext(pathState: .satisfied, route: .unavailable)
-        let gateway = RecordingDiagnosticGatewayMeasurer()
-
-        let result = await GatewayReachabilityCheck(
-            context: context,
-            gatewayMeasuring: gateway
-        ).run()
-
-        #expect(result.status == .indeterminate)
-        #expect(await gateway.targets.isEmpty)
-    }
-
-    @Test("blocked and skipped statuses have distinct presentations")
-    func blockedAndSkippedPresentation() {
-        #expect(NetworkDiagnosticStatus.blocked.presentation == .init(
-            labelKey: "network_diagnostics.status.blocked",
-            icon: "lock.fill",
-            tone: .informational
-        ))
-        #expect(NetworkDiagnosticStatus.skipped.presentation == .init(
-            labelKey: "network_diagnostics.status.skipped",
-            icon: "forward.fill",
-            tone: .muted
-        ))
-        #expect(NetworkDiagnosticStatus.blocked.presentation != NetworkDiagnosticStatus.indeterminate.presentation)
-        #expect(NetworkDiagnosticStatus.skipped.presentation != NetworkDiagnosticStatus.indeterminate.presentation)
-    }
-
-    @Test("proxy authentication evidence maps to a credential remediation")
-    func proxyAuthenticationRemediation() {
-        let result = NetworkDiagnosticResult(
-            id: .proxy,
-            status: .abnormal,
-            summary: "Proxy authentication required",
-            evidence: [.init(code: "proxy.authentication-required", value: "407")]
-        )
-
-        let remediation = NetworkDiagnosticRemediation.forResult(result)
-
-        #expect(remediation.actionKey == "network_diagnostics.remediation.proxy_authentication.action")
-        #expect(remediation.rerunKey == "network_diagnostics.remediation.rerun")
-    }
-
-    @Test("indeterminate remediation does not claim a confirmed cause")
-    func indeterminateRemediationIsTentative() {
-        let result = NetworkDiagnosticResult(
-            id: .dns,
-            status: .indeterminate,
-            summary: "DNS could not be determined"
-        )
-
-        let remediation = NetworkDiagnosticRemediation.forResult(result)
-
-        #expect(remediation.causeKey == "network_diagnostics.remediation.indeterminate.cause")
-        #expect(remediation.actionKey == "network_diagnostics.remediation.indeterminate.action")
-    }
-
-    @Test("proxy remediation follows final route facts instead of failed candidates")
-    func proxyRemediationUsesFinalRouteFacts() {
-        let result = NetworkDiagnosticResult(
-            id: .proxy,
-            status: .indeterminate,
-            summary: "proxy route recovered",
-            evidence: [.init(code: "proxy.authentication-required", value: "407")],
-            proxyFacts: .init(http: .available, https: .available)
-        )
-
-        #expect(NetworkDiagnosticRemediation.forResult(result).actionKey
-            == "network_diagnostics.remediation.indeterminate.action")
-    }
-
-    @Test("a blocked probe does not become an independent network fault")
-    func blockedProbeIsNotAbnormal() {
-        let result = NetworkDiagnosticResult.blocked(id: .internet, summary: "DNS is unavailable")
-        #expect(result.status == .blocked)
-        #expect(result.detail == "DNS is unavailable")
-    }
-
-    @Test("runner blocks gateway reachability after a hard path failure")
-    func runnerBlocksGatewayReachabilityAfterPathFailure() async {
-        let recorder = DiagnosticTestRecorder()
-        let checks: [any DiagnosticCheck] = [
-            StubDiagnosticCheck(id: .path, result: NetworkDiagnosticResult(id: .path, status: .abnormal, summary: "no path"), recorder: recorder),
-            StubDiagnosticCheck(id: .gatewayReachability, result: NetworkDiagnosticResult(id: .gatewayReachability, status: .normal, summary: "unused"), recorder: recorder),
-            StubDiagnosticCheck(id: .dns, result: NetworkDiagnosticResult(id: .dns, status: .normal, summary: "unused"), recorder: recorder),
-        ]
-        let runner = DiagnosticRunner(checks: checks)
-
-        let results = (await runner.run { _ in }).results
-
-        #expect(results.map(\.id) == [.path, .gatewayReachability, .dns])
-        #expect(results[1].status == .blocked)
-        #expect(results[1].evidence.contains(.init(code: "blocked.by", value: "path")))
-    }
-
-    @Test("gateway reachability abnormal does not block downstream checks")
-    func gatewayReachabilityAbnormalDoesNotBlockDownstream() async {
-        let recorder = DiagnosticTestRecorder()
-        let checks: [any DiagnosticCheck] = NetworkDiagnosticCheckID.allCases.map { id in
-            let status: NetworkDiagnosticStatus = id == .gatewayReachability ? .abnormal : .normal
-            return StubDiagnosticCheck(id: id, result: NetworkDiagnosticResult(id: id, status: status, summary: id.rawValue), recorder: recorder)
-        }
-        let runner = DiagnosticRunner(checks: checks)
-
-        let results = (await runner.run { _ in }).results
-
-        #expect(results.map(\.id) == NetworkDiagnosticCheckID.allCases)
-        #expect(results[1].status == .abnormal)
-        #expect(results[2].status == .normal)
-        #expect(results[3].status == .normal)
-        #expect(results[4].status == .normal)
-    }
-
-    @Test("gateway reachability abnormal drives the needs attention conclusion")
-    func gatewayReachabilityDrivesNeedsAttention() {
-        let results = [
-            NetworkDiagnosticResult(id: .path, status: .normal, summary: ""),
-            NetworkDiagnosticResult(id: .gatewayReachability, status: .abnormal, summary: ""),
-            NetworkDiagnosticResult(id: .dns, status: .normal, summary: ""),
-            NetworkDiagnosticResult(id: .internet, status: .normal, summary: ""),
-            NetworkDiagnosticResult(id: .ipv6, status: .skipped, summary: ""),
-            NetworkDiagnosticResult(id: .proxy, status: .normal, summary: ""),
-        ]
-
-        #expect(NetworkDiagnosticConclusion.evaluate(results) == .needsAttention)
-        #expect(NetworkDiagnosticConclusion.primaryIssue(in: results)?.id == .gatewayReachability)
-    }
-
-    @Test("stage resolver aggregates this mac from path dns and proxy and maps lan to its check")
-    func stageResolverThisMacAndLan() {
-        let resolver = NetworkDiagnosticStageResolver()
-        let abnormalDNS: [NetworkDiagnosticCheckID: NetworkDiagnosticResult] = [
-            .path: .init(id: .path, status: .normal, summary: ""),
-            .dns: .init(id: .dns, status: .abnormal, summary: ""),
-            .proxy: .init(id: .proxy, status: .normal, summary: ""),
-            .gatewayReachability: .init(id: .gatewayReachability, status: .normal, summary: ""),
-        ]
-
-        #expect(resolver.status(for: .thisMac, results: abnormalDNS) == .abnormal)
-        #expect(resolver.status(for: .lan, results: abnormalDNS) == .normal)
-        let proxyOnly: [NetworkDiagnosticCheckID: NetworkDiagnosticResult] = [
-            .path: .init(id: .path, status: .normal, summary: ""),
-            .dns: .init(id: .dns, status: .normal, summary: ""),
-            .proxy: .init(id: .proxy, status: .abnormal, summary: ""),
-            .gatewayReachability: .init(id: .gatewayReachability, status: .normal, summary: ""),
-        ]
-        #expect(resolver.status(for: .thisMac, results: proxyOnly) == .abnormal)
-        #expect(resolver.status(for: .thisMac, results: [:]) == nil)
-        let missingContributors: [NetworkDiagnosticCheckID: NetworkDiagnosticResult] = [
-            .path: .init(id: .path, status: .normal, summary: ""),
-        ]
-        #expect(resolver.status(for: .thisMac, results: missingContributors) == nil)
-    }
-
-    @Test("internet stage follows the internet check and ignores dns and ipv6")
-    func stageResolverInternet() {
-        let resolver = NetworkDiagnosticStageResolver()
-
-        #expect(resolver.status(for: .internet, results: [.internet: .init(id: .internet, status: .normal, summary: "")]) == .normal)
-        #expect(resolver.status(for: .internet, results: [.internet: .init(id: .internet, status: .abnormal, summary: "")]) == .abnormal)
-        #expect(resolver.status(for: .internet, results: [.internet: .init(id: .internet, status: .indeterminate, summary: "")]) == .indeterminate)
-        #expect(resolver.status(for: .internet, results: [.internet: .init(id: .internet, status: .blocked, summary: "")]) == .blocked)
-        let extras: [NetworkDiagnosticCheckID: NetworkDiagnosticResult] = [
-            .internet: .init(id: .internet, status: .normal, summary: ""),
-            .dns: .init(id: .dns, status: .abnormal, summary: ""),
-            .ipv6: .init(id: .ipv6, status: .abnormal, summary: ""),
-        ]
-        #expect(resolver.status(for: .internet, results: extras) == .normal)
-    }
-
-    @Test("internet stage requires the internet result")
-    func stageResolverRequiresInternetResult() {
-        let resolver = NetworkDiagnosticStageResolver()
-        let partial: [NetworkDiagnosticCheckID: NetworkDiagnosticResult] = [
-            .dns: .init(id: .dns, status: .normal, summary: ""),
-        ]
-
-        #expect(resolver.status(for: .internet, results: partial) == nil)
-    }
-
-    @Test("pipeline presentation maps stages to stations and edges")
-    func pipelinePresentation() {
-        let results: [NetworkDiagnosticCheckID: NetworkDiagnosticResult] = [
-            .path: .init(id: .path, status: .normal, summary: ""),
-            .gatewayReachability: .init(id: .gatewayReachability, status: .abnormal, summary: ""),
-            .dns: .init(id: .dns, status: .normal, summary: ""),
-            .internet: .init(id: .internet, status: .normal, summary: ""),
-            .ipv6: .init(id: .ipv6, status: .skipped, summary: ""),
-            .proxy: .init(id: .proxy, status: .normal, summary: ""),
-        ]
-        let presentation = NetworkDiagnosticsPipelinePresentation.from(
-            results: results,
-            executionPhases: [:]
-        )
-
-        #expect(presentation.stations.map(\.kind) == [.thisMac, .router, .internet])
-        #expect(presentation.edges.map(\.kind) == [.lan, .internet])
-        #expect(presentation.stations[0].status == .normal)
-        #expect(presentation.stations[1].status == .abnormal)
-        #expect(presentation.stations[1].isUnreachable)
-        #expect(presentation.stations[2].status == .normal)
-        #expect(!presentation.stations[2].isUnreachable)
-        #expect(presentation.edges[0].status == .abnormal)
-        #expect(presentation.edges[1].status == .normal)
-    }
-
-    @Test("pipeline edge becomes active while its checks are running")
-    func pipelineEdgeActive() {
-        let executionPhases: [NetworkDiagnosticCheckID: NetworkDiagnosticExecutionPhase] = [
-            .gatewayReachability: .checking,
-        ]
-        let presentation = NetworkDiagnosticsPipelinePresentation.from(
-            results: [:],
-            executionPhases: executionPhases
-        )
-
-        #expect(presentation.edges[0].isActive)
-        #expect(!presentation.edges[1].isActive)
-    }
-
-    @Test("runner executes checks and publishes results in order")
-    func runnerOrder() async {
-        let invocations = DiagnosticTestRecorder()
-        let publications = DiagnosticTestRecorder()
-        let checks: [any DiagnosticCheck] = NetworkDiagnosticCheckID.allCases.map { id in
-            StubDiagnosticCheck(
-                id: id,
-                result: NetworkDiagnosticResult(id: id, status: .normal, summary: id.rawValue),
-                recorder: invocations
-            )
-        }
-
-        let results = (await DiagnosticRunner(checks: checks).run { result in
-            await publications.record(result.id)
-        }).results
-
-        #expect(results.map(\.id) == NetworkDiagnosticCheckID.allCases)
-        #expect(await invocations.values == NetworkDiagnosticCheckID.allCases)
-        #expect(await publications.values == NetworkDiagnosticCheckID.allCases)
-    }
-
-    @Test("runner enforces an overall session budget")
-    func runnerEnforcesOverallBudget() async {
-        let probe = BudgetAwareDiagnosticProbe()
-        let clock = ManualDiagnosticClock()
-        let task = Task {
-            await DiagnosticRunner(
-                checks: [BudgetAwareDiagnosticCheck(probe: probe)],
-                sessionBudget: .seconds(1),
-                clock: clock
-            ).run { _ in }
-        }
-
-        await probe.waitForInvocation()
-        await clock.waitForSleeperCount(1)
-        await clock.set(clock.origin.advanced(by: .seconds(1)))
-
-        let outcome = await task.value
-
-        #expect(await probe.wasCancelled)
-        #expect(outcome.results.map(\.id) == [.path])
-        #expect(outcome.results.first?.status == .indeterminate)
-        #expect(outcome.results.first?.evidence == [.init(code: "check.timeout", value: nil)])
-        #expect(outcome.pendingIDs.isEmpty)
-        #expect(outcome.endReason == .timedOut)
-    }
-
-    @Test("runner returns when a cancelled check ignores cancellation")
-    func runnerDoesNotWaitForCancellationIgnoringCheck() async {
-        let probe = CancellationIgnoringDiagnosticProbe()
-        let task = Task {
-            await DiagnosticRunner(
-                checks: [CancellationIgnoringProbeDiagnosticCheck(probe: probe)],
-                sessionBudget: .seconds(30)
-            ).run { _ in }
-        }
-
-        await probe.waitForInvocationCount(1)
-        task.cancel()
-        let outcome = await task.value
-
-        #expect(outcome.results.isEmpty)
-        #expect(outcome.pendingIDs == [.path])
-        #expect(outcome.endReason == .cancelled)
-        await probe.release(invocation: 1)
-    }
-
-    @Test("publication gate rejects an old run after a replacement starts")
-    func oldRunCannotPublishIntoNewRun() {
-        let oldID = UUID()
-        let newID = UUID()
-        let gate = DiagnosticPublicationGate(activeRunID: newID)
-
-        #expect(!gate.accepts(oldID))
-        #expect(gate.accepts(newID))
-    }
-
-    @Test("runner preserves completed results and identifies pending checks on cancellation")
-    func runnerReturnsPartialCancelledOutcome() async {
-        let probe = BlockingDiagnosticProbe()
-        let checks: [any DiagnosticCheck] = [
-            StubDiagnosticCheck(
-                id: .path,
-                result: .init(id: .path, status: .normal, summary: "path"),
-                recorder: DiagnosticTestRecorder()
-            ),
-            BlockingProbeDiagnosticCheck(id: .gatewayReachability, probe: probe),
-        ]
-        let runner = DiagnosticRunner(checks: checks)
-        let task = Task { await runner.run { _ in } }
-
-        await probe.waitForInvocationCount(1)
-        task.cancel()
-        let outcome = await task.value
-
-        #expect(outcome.results.map(\.id) == [.path])
-        #expect(outcome.pendingIDs == [.gatewayReachability])
-        #expect(outcome.endReason == .cancelled)
-    }
-
-    @Test("indeterminate DNS does not block internet or IPv6")
-    func indeterminateDNSDoesNotBlockInternetOrIPv6() async {
-        let invocations = DiagnosticTestRecorder()
-        let checks: [any DiagnosticCheck] = NetworkDiagnosticCheckID.allCases.map { id in
-            StubDiagnosticCheck(
-                id: id,
-                result: NetworkDiagnosticResult(
-                    id: id,
-                    status: id == .dns ? .indeterminate : .normal,
-                    summary: id.rawValue
-                ),
-                recorder: invocations
-            )
-        }
-
-        let results = (await DiagnosticRunner(checks: checks).run { _ in }).results
-
-        #expect(results.map(\.status) == [.normal, .normal, .indeterminate, .normal, .normal, .normal])
-        #expect(await invocations.values == [.path, .gatewayReachability, .dns, .internet, .ipv6, .proxy])
-    }
-
-    @Test("abnormal DNS does not block independent hostname checks")
-    func abnormalDNSDoesNotBlockHostnameDependentChecks() async {
-        let invocations = DiagnosticTestRecorder()
-        let checks: [any DiagnosticCheck] = NetworkDiagnosticCheckID.allCases.map { id in
-            StubDiagnosticCheck(
-                id: id,
-                result: NetworkDiagnosticResult(
-                    id: id,
-                    status: id == .dns ? .abnormal : .normal,
-                    summary: id.rawValue
-                ),
-                recorder: invocations
-            )
-        }
-
-        let results = (await DiagnosticRunner(checks: checks).run { _ in }).results
-
-        #expect(results.map(\.status) == [.normal, .normal, .abnormal, .normal, .normal, .normal])
-        #expect(await invocations.values == [.path, .gatewayReachability, .dns, .internet, .ipv6, .proxy])
-    }
-
-    @Test("indeterminate path continues independent evidence probes")
-    func indeterminatePathContinuesIndependentEvidenceProbes() async {
-        let invocations = DiagnosticTestRecorder()
-        let checks: [any DiagnosticCheck] = NetworkDiagnosticCheckID.allCases.map { id in
-            StubDiagnosticCheck(
-                id: id,
-                result: NetworkDiagnosticResult(
-                    id: id,
-                    status: id == .path ? .indeterminate : .normal,
-                    summary: id.rawValue
-                ),
-                recorder: invocations
-            )
-        }
-
-        let results = (await DiagnosticRunner(checks: checks).run { _ in }).results
-
-        #expect(results.map(\.status) == [.indeterminate, .normal, .normal, .normal, .normal, .normal])
-        #expect(await invocations.values == [.path, .gatewayReachability, .dns, .internet, .ipv6, .proxy])
-    }
-
-    @Test("abnormal path blocks network probes")
-    func abnormalPathBlocksNetworkProbes() async {
-        let invocations = DiagnosticTestRecorder()
-        let checks: [any DiagnosticCheck] = NetworkDiagnosticCheckID.allCases.map { id in
-            StubDiagnosticCheck(
-                id: id,
-                result: NetworkDiagnosticResult(
-                    id: id,
-                    status: id == .path ? .abnormal : .normal,
-                    summary: id.rawValue
-                ),
-                recorder: invocations
-            )
-        }
-
-        let results = (await DiagnosticRunner(checks: checks).run { _ in }).results
-
-        #expect(results.map(\.status) == [.abnormal, .blocked, .blocked, .blocked, .blocked, .blocked])
-        #expect(await invocations.values == [.path])
-    }
-
-    @Test("system path failure makes the network unavailable")
-    func unavailableConclusion() {
-        let results = makeResults(path: .abnormal, dns: .normal, internet: .normal, proxy: .normal)
-        #expect(NetworkDiagnosticConclusion.evaluate(results) == .networkUnavailable)
-    }
-
-    @Test("DNS or proxy failure needs attention")
-    func abnormalConclusion() {
-        #expect(NetworkDiagnosticConclusion.evaluate(
-            makeResults(path: .normal, dns: .abnormal, internet: .normal, proxy: .normal)
-        ) == .needsAttention)
-        #expect(NetworkDiagnosticConclusion.evaluate(
-            makeResults(path: .normal, dns: .normal, internet: .normal, proxy: .abnormal)
-        ) == .needsAttention)
-    }
-
-    @Test("failed base HTTPS suppresses proxy endpoint symptom without declaring outage")
-    func conclusionSuppressesProxySymptom() {
-        var results = makeResults(
-            path: .normal,
-            dns: .indeterminate,
-            internet: .abnormal,
-            proxy: .abnormal
-        )
-        results[3] = NetworkDiagnosticResult(
-            id: .internet,
-            status: .abnormal,
-            summary: "internet",
-            evidence: [.init(
-                code: "https.connectivity-error",
-                value: String(URLError.timedOut.rawValue)
-            )]
-        )
-
-        #expect(NetworkDiagnosticConclusion.evaluate(results) == .needsAttention)
-    }
-
-    @Test("failed direct path with a usable HTTPS proxy needs attention")
-    func usableProxyPreventsNetworkUnavailableConclusion() {
-        var results = makeResults(
-            path: .normal,
-            dns: .normal,
-            internet: .abnormal,
-            proxy: .normal
-        )
-        results[5] = NetworkDiagnosticResult(
-            id: .proxy,
-            status: .normal,
-            summary: "proxy",
-            evidence: [.init(code: "proxy.https.egress-status", value: "200")]
-        )
-
-        #expect(NetworkDiagnosticConclusion.evaluate(results) == .needsAttention)
-    }
-
-    @Test("DIRECT routing without direct egress needs attention")
-    func directRouteDecisionDoesNotProveConnectivity() {
-        var results = makeResults(
-            path: .normal,
-            dns: .normal,
-            internet: .abnormal,
-            proxy: .normal
-        )
-        results[3] = NetworkDiagnosticResult(
-            id: .internet,
-            status: .abnormal,
-            summary: "internet",
-            evidence: [.init(
-                code: "https.connectivity-error",
-                value: String(URLError.notConnectedToInternet.rawValue)
-            )]
-        )
-        results[5] = NetworkDiagnosticResult(
-            id: .proxy,
-            status: .normal,
-            summary: "proxy",
-            evidence: [.init(code: "proxy.https.egress-status", value: "base-check")]
-        )
-
-        #expect(NetworkDiagnosticConclusion.evaluate(results) == .needsAttention)
-    }
-
-    @Test("base HTTPS success and explicit proxy failure needs attention")
-    func proxyIsIndependentWhenBasePathWorks() {
-        let results: [NetworkDiagnosticResult] = makeResults(
-            path: .normal,
-            dns: .normal,
-            internet: .normal,
-            proxy: .abnormal
-        )
-
-        #expect(NetworkDiagnosticConclusion.evaluate(results) == .needsAttention)
-    }
-
-    @Test("captive portal symptom after HTTPS success needs attention")
-    func captivePortalSymptomAfterHTTPSSuccess() {
-        let results = [
-            NetworkDiagnosticResult(id: .path, status: .normal, summary: "path"),
-            NetworkDiagnosticResult(id: .gatewayReachability, status: .normal, summary: "gateway"),
-            NetworkDiagnosticResult(id: .dns, status: .normal, summary: "dns"),
-            NetworkDiagnosticResult(
-                id: .internet,
-                status: .abnormal,
-                summary: "internet",
-                evidence: [
-                    .init(code: "https.available", value: "200"),
-                    .init(code: "captive-portal.suspected", value: nil),
-                ]
-            ),
-            NetworkDiagnosticResult(id: .ipv6, status: .skipped, summary: "ipv6"),
-            NetworkDiagnosticResult(id: .proxy, status: .normal, summary: "proxy"),
-        ]
-
-        #expect(NetworkDiagnosticConclusion.evaluate(results) == .needsAttention)
-        let assessment = NetworkDiagnosticAssessmentResolver().resolve(
-            results: Dictionary(uniqueKeysWithValues: results.map { ($0.id, $0) }),
-            complete: true
-        )
-        #expect(assessment.primaryIssue?.id == .internet)
-    }
-
-    @Test("an indeterminate result needs attention")
-    func indeterminateConclusion() {
-        let results = makeResults(path: .normal, dns: .indeterminate, internet: .normal, proxy: .normal)
-        #expect(NetworkDiagnosticConclusion.evaluate(results) == .needsAttention)
-    }
-
-    @Test("normal required results and skipped optional IPv6 make the network normal")
-    func fourResultNormalConclusion() {
-        let results = makeResults(path: .normal, dns: .normal, internet: .normal, proxy: .normal)
-        #expect(NetworkDiagnosticConclusion.evaluate(results) == .networkNormal)
-    }
-
-    @Test("verified base connectivity keeps an unverified DIRECT proxy route neutral")
-    func verifiedBaseConnectivityKeepsDirectRouteNeutral() {
-        var results = makeResults(
-            path: .normal,
-            dns: .normal,
-            internet: .normal,
-            proxy: .indeterminate
-        )
-        results[5] = NetworkDiagnosticResult(
-            id: .proxy,
-            status: .indeterminate,
-            summary: "direct routing selected",
-            evidence: [.init(code: "proxy.https.egress-status", value: "base-check")],
-            proxyFacts: .init(http: .unverified, https: .unverified)
-        )
-
-        #expect(NetworkDiagnosticConclusion.evaluate(results) == .networkNormal)
-    }
-
-    @Test("PAC resolution failures are not neutralized as explicit DIRECT")
-    func pacResolutionFailureDoesNotBecomeNormal() {
-        let results = makeResults(
-            path: .normal,
-            dns: .normal,
-            internet: .normal,
-            proxy: .indeterminate
-        ).map { result in
-            guard result.id == .proxy else { return result }
-            return NetworkDiagnosticResult(
-                id: .proxy,
-                status: .indeterminate,
-                summary: "proxy resolution failed",
-                evidence: [
-                    .init(code: "proxy.http.resolution", value: "pac-timeout"),
-                    .init(code: "proxy.https.resolution", value: "pac-timeout"),
-                ],
-                proxyFacts: .init(http: .unverified, https: .unverified)
-            )
-        }
-
-        let assessment = NetworkDiagnosticAssessmentResolver().resolve(
-            results: Dictionary(uniqueKeysWithValues: results.map { ($0.id, $0) }),
-            complete: true
-        )
-
-        #expect(assessment.conclusion == .needsAttention)
-    }
-
-    @Test("a single HTTPS target failure is needs attention, not universal outage")
-    func singleHTTPSFailureIsNotNetworkUnavailable() {
-        var results = makeResults(
-            path: .normal,
-            dns: .normal,
-            internet: .abnormal,
-            proxy: .indeterminate
-        )
-        results[3] = NetworkDiagnosticResult(
-            id: .internet,
-            status: .abnormal,
-            summary: "one HTTPS target failed",
-            evidence: [.init(code: "https.connectivity-error", value: nil)]
-        )
-
-        #expect(NetworkDiagnosticConclusion.evaluate(results) == .needsAttention)
-    }
-
-    @Test("DNS uncertainty is not hidden by a neutral DIRECT proxy")
-    func dnsUncertaintyRemainsVisibleAlongsideDirectProxy() {
-        var results = makeResults(
-            path: .normal,
-            dns: .indeterminate,
-            internet: .normal,
-            proxy: .indeterminate
-        )
-        results[5] = NetworkDiagnosticResult(
-            id: .proxy,
-            status: .indeterminate,
-            summary: "direct routing selected",
-            proxyFacts: .init(http: .unverified, https: .unverified)
-        )
-
-        let assessment = NetworkDiagnosticAssessmentResolver().resolve(
-            results: Dictionary(uniqueKeysWithValues: results.map { ($0.id, $0) }),
-            complete: true
-        )
-
-        #expect(assessment.conclusion == .needsAttention)
-        #expect(assessment.stages.first { $0.stage == .thisMac }?.status == .indeterminate)
-    }
-
-    @Test("HTTPS availability is retained even when HTTP proxy routing fails")
-    func httpsAvailabilityDoesNotDependOnHTTP() {
-        let facts = DiagnosticProxyFacts(http: .unavailable, https: .available)
-        #expect(facts.https == .available)
-
-        let results: [NetworkDiagnosticResult] = makeResults(
-            path: .normal,
-            dns: .normal,
-            internet: .abnormal,
-            proxy: .abnormal
-        ).map { result in
-            if result.id == .internet {
-                return NetworkDiagnosticResult(
-                    id: result.id,
-                    status: result.status,
-                    summary: result.summary,
-                    evidence: [.init(code: "https.connectivity-error", value: nil)]
-                )
-            }
-            if result.id == .proxy {
-                return NetworkDiagnosticResult(
-                    id: result.id,
-                    status: result.status,
-                    summary: result.summary,
-                    evidence: [.init(code: "proxy.endpoint-unavailable", value: nil)],
-                    proxyFacts: facts
-                )
-            }
-            return result
-        }
-
-        let assessment = NetworkDiagnosticAssessmentResolver().resolve(
-            results: Dictionary(uniqueKeysWithValues: results.map { ($0.id, $0) }),
-            complete: true
-        )
-        #expect(assessment.conclusion == .needsAttention)
-        #expect(assessment.primaryIssue?.id == .proxy)
-    }
-
-    @Test("recovered proxy candidates do not leave an authentication action")
-    func recoveredProxyCandidateDoesNotRemainActionable() {
-        let results: [NetworkDiagnosticResult] = makeResults(
-            path: .normal,
-            dns: .normal,
-            internet: .normal,
-            proxy: .normal
-        ).map { result in
-            guard result.id == .proxy else { return result }
-            return NetworkDiagnosticResult(
-                id: .proxy,
-                status: .normal,
-                summary: "HTTPS proxy recovered",
-                evidence: [.init(code: "proxy.https.authentication-required", value: "407")],
-                proxyFacts: .init(http: .available, https: .available)
-            )
-        }
-
-        let assessment = NetworkDiagnosticAssessmentResolver().resolve(
-            results: Dictionary(uniqueKeysWithValues: results.map { ($0.id, $0) }),
-            complete: true
-        )
-        #expect(assessment.conclusion == .networkNormal)
-        #expect(assessment.primaryIssue == nil)
-    }
-
-    @Test("no global IPv6 address is skipped without changing a normal conclusion")
-    func ipv6SkippedIsConclusionNeutral() async {
-        let ipv6 = await IPv6ControlEndpointCheck(
-            loader: StubIPv6Loader(.noGlobalAddress)
-        ).run()
-
-        #expect(ipv6.id == .ipv6)
-        #expect(ipv6.status == .skipped)
-        #expect(NetworkDiagnosticConclusion.evaluate(
-            makeResults(
-                path: .normal,
-                dns: .normal,
-                internet: .normal,
-                ipv6: ipv6.status,
-                proxy: .normal
-            )
-        ) == .networkNormal)
-    }
-
-    @Test("IPv6 failure does not affect the overall conclusion")
-    func ipv6IsConclusionNeutral() async {
-        let ipv6 = await IPv6ControlEndpointCheck(loader: StubIPv6Loader(.failed)).run()
-
-        #expect(ipv6.status == .indeterminate)
-        #expect(ipv6.evidence.contains(.init(code: "ipv6.unavailable", value: nil)))
-        #expect(NetworkDiagnosticConclusion.evaluate(
-            makeResults(
-                path: .normal,
-                dns: .normal,
-                internet: .normal,
-                ipv6: ipv6.status,
-                proxy: .normal
-            )
-        ) == .networkNormal)
-    }
-
-    @Test("IPv6 control endpoint success is normal")
-    func ipv6Success() async {
-        let result = await IPv6ControlEndpointCheck(loader: StubIPv6Loader(.succeeded)).run()
-
-        #expect(result.status == .normal)
-        #expect(result.evidence.contains(.init(code: "ipv6.available", value: nil)))
-    }
-
-    @Test("system IPv6 loader connects only to a resolved IPv6 literal")
-    func ipv6LoaderForcesResolvedAddress() async {
-        let recorder = IPv6LoaderTestRecorder()
-        let loader = SystemIPv6ControlEndpointLoader(
-            addressSource: StubGlobalIPv6AddressSource(hasAddress: true),
-            resolver: StubIPv6AddressResolver(addresses: ["2001:db8::42"]),
-            connector: RecordingIPv6HTTPSConnector(succeeds: true, recorder: recorder)
-        )
-        let endpoint = URL(string: "https://control.example/health")!
-
-        let outcome = await loader.load(url: endpoint, timeout: .seconds(4))
-
-        #expect(outcome == .succeeded)
-        let requests = await recorder.requests
-        #expect(requests.count == 1)
-        #expect(requests.first?.url == endpoint)
-        #expect(requests.first?.ipv6Address == "2001:db8::42")
-        #expect(requests.first?.serverName == "control.example")
-        #expect(requests.first.map { $0.timeout > .zero && $0.timeout <= .seconds(4) } == true)
-    }
-
-    @Test("system IPv6 loader tries later resolved addresses")
-    func ipv6LoaderFallsBackAcrossAddresses() async {
-        let connector = SequencedIPv6HTTPSConnector(successfulAddress: "2001:db8::2")
-        let loader = SystemIPv6ControlEndpointLoader(
-            addressSource: StubGlobalIPv6AddressSource(hasAddress: true),
-            resolver: StubIPv6AddressResolver(addresses: ["2001:db8::1", "2001:db8::2"]),
-            connector: connector
-        )
-
-        let outcome = await loader.load(
-            url: URL(string: "https://control.example/health")!,
-            timeout: .seconds(4)
-        )
-
-        #expect(outcome == .succeeded)
-        #expect(await connector.addresses == ["2001:db8::1", "2001:db8::2"])
-        let timeouts = await connector.timeouts
-        #expect(timeouts.count == 2)
-        #expect(timeouts.first.map { $0 > .zero && $0 <= .seconds(2) } == true)
-        #expect(timeouts.last.map { $0 > .zero && $0 <= .seconds(4) } == true)
-    }
-
-    @Test("system IPv6 loader skips resolution when no global address is present")
-    func ipv6LoaderSkipsWithoutGlobalAddress() async {
-        let resolver = RecordingIPv6AddressResolver(addresses: ["2001:db8::42"])
-        let loader = SystemIPv6ControlEndpointLoader(
-            addressSource: StubGlobalIPv6AddressSource(hasAddress: false),
-            resolver: resolver,
-            connector: StubIPv6HTTPSConnector(succeeds: true)
-        )
-
-        let outcome = await loader.load(
-            url: URL(string: "https://control.example/")!,
-            timeout: .seconds(4)
-        )
-
-        #expect(outcome == .noGlobalAddress)
-        #expect(await resolver.hosts.isEmpty)
-    }
-
-    @Test("IPv6 DNS configures delivery before the resolver context owns the service")
-    func ipv6DNSActivationOrder() {
-        var events: [String] = []
-
-        IPv6DNSServiceActivation.configureThenInstall(
-            configureDelivery: { events.append("configure") },
-            installOwnership: { events.append("install") }
-        )
-
-        #expect(events == ["configure", "install"])
-    }
-
-    @Test("an incomplete run has no conclusion")
-    func incompleteConclusion() {
-        let result = NetworkDiagnosticResult(id: .path, status: .normal, summary: "ok")
-        #expect(NetworkDiagnosticConclusion.evaluate([result]) == nil)
-    }
-
-    @Test("workbench table adapts columns to available width")
-    func adaptiveWorkbenchLayoutMode() {
-        #expect(NetworkDiagnosticsWorkbenchLayout.mode(for: 519) == .compact)
-        #expect(NetworkDiagnosticsWorkbenchLayout.mode(for: 520) == .condensed)
-        #expect(NetworkDiagnosticsWorkbenchLayout.mode(for: 719) == .condensed)
-        #expect(NetworkDiagnosticsWorkbenchLayout.mode(for: 720) == .regular)
-    }
-
-    @Test("result table uses comfortable rows without alternating empty backgrounds")
-    func comfortableResultTablePresentation() {
-        #expect(NetworkDiagnosticsTablePresentation.minimumRowHeight == 54)
-        #expect(NetworkDiagnosticsTablePresentation.usesAlternatingRowBackgrounds == false)
-    }
-
-    @Test("workbench items interleave stage headers with check rows")
-    func workbenchItemGrouping() {
-        let path = NetworkDiagnosticResult(
-            id: .path,
-            status: .normal,
-            summary: "connected"
-        )
-        let executionPhases: [NetworkDiagnosticCheckID: NetworkDiagnosticExecutionPhase] = [
-            .path: .completed,
-            .dns: .checking,
-            .proxy: .waiting,
-        ]
-
-        #expect(NetworkDiagnosticsPresentation.workbenchItems(
-            pagePhase: .idle,
-            executionPhases: executionPhases,
-            results: [:]
-        ).isEmpty)
-
-        let runningItems = NetworkDiagnosticsPresentation.workbenchItems(
-            pagePhase: .running,
-            executionPhases: executionPhases,
-            results: [.path: path],
-            checkIDs: [.path, .dns, .proxy]
-        )
-        #expect(runningItems.map(\.id) == ["header.thisMac", "check.path", "check.dns"])
-        #expect(runningItems[0] == .stageHeader(.thisMac))
-        #expect(runningItems[1] == .check(.init(id: .path, executionPhase: .completed, result: path)))
-
-        let completedResults = Dictionary(uniqueKeysWithValues: makeResults(
-            path: .normal,
-            dns: .abnormal,
-            internet: .normal,
-            proxy: .indeterminate
-        ).map { ($0.id, $0) })
-        let completedItems = NetworkDiagnosticsPresentation.workbenchItems(
-            pagePhase: .completed,
-            executionPhases: executionPhases,
-            results: completedResults,
-            checkIDs: [.path, .dns, .ipv6, .proxy]
-        )
-        #expect(completedItems.map(\.id) == [
-            "header.thisMac", "check.path", "check.dns", "check.proxy",
-            "header.additional", "check.ipv6",
-        ])
-        #expect(completedItems[4] == .additionalHeader)
-
-        let timedOutItems = NetworkDiagnosticsPresentation.workbenchItems(
-            pagePhase: .completed,
-            executionPhases: [.path: .completed, .dns: .waiting],
-            results: [.path: path],
-            checkIDs: [.path, .dns],
-            endReason: .timedOut
-        )
-        #expect(timedOutItems.last == .check(.init(
-            id: .dns,
-            executionPhase: .waiting,
-            result: nil,
-            pendingReason: .timedOut
-        )))
-    }
-
-    @Test("system path check maps path states")
-    func pathMapping() async {
-        let satisfied = await NetworkConnectivityCheck(pathSource: StubPathSource(.satisfied)).run()
-        #expect(satisfied.id == .path)
-        #expect(satisfied.status == .normal)
-        #expect(await NetworkConnectivityCheck(pathSource: StubPathSource(.unsatisfied)).run().status == .abnormal)
-        #expect(await NetworkConnectivityCheck(pathSource: StubPathSource(.requiresConnection)).run().status == .indeterminate)
-        #expect(await NetworkConnectivityCheck(pathSource: StubPathSource(nil)).run().status == .indeterminate)
-    }
-
-    @Test("a satisfied system path is not an internet-success result")
-    func pathCheckReportsPathOnly() async {
-        let result = await NetworkConnectivityCheck(pathSource: StubPathSource(.satisfied)).run()
-
-        #expect(result.id == .path)
-        #expect(result.id != .internet)
-        #expect(result.detail == String(
-            localized: "network_diagnostics.path.normal.summary",
-            comment: "Network self-check system path success detail"
-        ))
-    }
-
-    @Test("gateway reachability check succeeds with latency evidence")
-    func gatewayReachabilitySuccess() async {
-        let result = await GatewayReachabilityCheck(
-            interfaceSource: StubNetworkInterfaceSource(interface: makeNetworkInterface(router: "192.0.2.1")),
-            gatewayLatency: StubGatewayLatencyProvider(result: GatewayLatencyResult(
-                timestamp: Date(),
-                routerIP: "192.0.2.1",
-                latencyMs: 2.5
-            ))
-        ).run()
-
-        #expect(result.id == .gatewayReachability)
-        #expect(result.status == .normal)
-        #expect(result.evidence.contains(.init(code: "gateway.latency-ms", value: "2.5")))
-    }
-
-    @Test("gateway nonresponse remains indeterminate")
-    func gatewayNonresponseIsIndeterminate() async {
-        let result = await GatewayReachabilityCheck(
-            interfaceSource: StubNetworkInterfaceSource(interface: makeNetworkInterface(router: "192.0.2.1")),
-            gatewayLatency: StubGatewayLatencyProvider(result: GatewayLatencyResult(
-                timestamp: Date(),
-                routerIP: "192.0.2.1",
-                error: .gatewayPingFailed("192.0.2.1")
-            ))
-        ).run()
-
-        #expect(result.status == .indeterminate)
-        #expect(result.evidence.contains(.init(code: "gateway.no-response", value: "192.0.2.1")))
-    }
-
-    @Test("successful HTTPS access neutralizes gateway ICMP nonresponse")
-    func gatewayNonresponseDoesNotDowngradeSuccessfulHTTPS() {
-        var results = makeResults(
-            path: .normal,
-            gateway: .indeterminate,
-            dns: .normal,
-            internet: .normal,
-            proxy: .normal
-        )
-        results[1] = NetworkDiagnosticResult(
-            id: .gatewayReachability,
-            status: .indeterminate,
-            summary: "gateway did not answer ICMP",
-            evidence: [.init(code: "gateway.no-response", value: "192.0.2.1")]
-        )
-        results[3] = NetworkDiagnosticResult(
-            id: .internet,
-            status: .normal,
-            summary: "HTTPS available",
-            evidence: [.init(code: "https.available", value: "200")]
-        )
-
-        let assessment = NetworkDiagnosticAssessmentResolver().resolve(
-            results: Dictionary(uniqueKeysWithValues: results.map { ($0.id, $0) }),
-            complete: true
-        )
-
-        #expect(assessment.conclusion == .networkNormal)
-        #expect(assessment.primaryIssue == nil)
-        #expect(assessment.stages.first { $0.stage == .lan }?.status == .indeterminate)
-    }
-
-    @Test("gateway reachability is indeterminate without a router IP")
-    func gatewayReachabilityWithoutRouterIP() async {
-        let result = await GatewayReachabilityCheck(
-            interfaceSource: StubNetworkInterfaceSource(interface: makeNetworkInterface(router: nil)),
-            gatewayLatency: StubGatewayLatencyProvider(result: GatewayLatencyResult(
-                timestamp: Date(),
-                error: .missingRouterIP
-            ))
-        ).run()
-
-        #expect(result.status == .indeterminate)
-        #expect(result.evidence.contains(.init(code: "gateway.unavailable", value: nil)))
-    }
-
-    @Test("path check keeps interface evidence without gateway latency")
-    func pathCheckKeepsInterfaceEvidence() async {
-        let result = await NetworkConnectivityCheck(
-            pathSource: StubPathSource(.satisfied),
-            interfaceSource: StubNetworkInterfaceSource(interface: makeNetworkInterface(router: "192.0.2.1"))
-        ).run()
-
-        #expect(result.status == .normal)
-        #expect(result.evidence.contains(.init(code: "path.interface", value: "en0")))
-        #expect(result.evidence.contains(.init(code: "path.local-ip", value: "192.0.2.10")))
-        #expect(result.evidence.contains(.init(code: "path.router", value: "192.0.2.1")))
-        #expect(!result.evidence.contains { $0.code.hasPrefix("gateway.") })
-    }
-
-    @Test("captive-portal probe rejects a rewritten response")
-    func captivePortalProbeRejectsRewrittenResponse() async {
-        let check = HTTPSControlEndpointCheck(
-            loader: StubControlLoader(
-                httpsStatus: 204,
-                httpStatus: 200,
-                httpBody: "login"
-            )
-        )
-
-        let result = await check.run()
-
-        #expect(result.id == .internet)
-        #expect(result.evidence.contains(.init(code: "captive-portal.suspected", value: nil)))
-    }
-
-    @Test("control endpoint success verifies HTTPS and Apple's captive-portal Success body")
-    func controlEndpointSuccess() async {
-        let check = HTTPSControlEndpointCheck(
-            loader: StubControlLoader(
-                httpsStatus: 200,
-                httpStatus: 200,
-                httpBody: "Success"
-            )
-        )
-
-        let result = await check.run()
-
-        #expect(result.status == .normal)
-        #expect(result.evidence.contains(.init(code: "https.available", value: "200")))
-        #expect(result.evidence.contains(.init(code: "captive-portal.clear", value: nil)))
-    }
-
-    @Test("independent control endpoints load concurrently and preserve evidence order")
-    func controlEndpointsRunConcurrently() async {
-        let loader = ConcurrentControlLoader()
-        let result = await HTTPSControlEndpointCheck(loader: loader).run()
-
-        #expect(await loader.maximumInFlight == 2)
-        #expect(result.evidence.map(\.code).prefix(2) == ["https.available", "captive-portal.clear"])
-    }
-
-    @Test("control endpoint metrics are redacted to bounded transaction fields")
-    func controlEndpointMetricsAreRedacted() async {
-        let result = await HTTPSControlEndpointCheck(
-            loader: MetricsControlLoader(
-                metrics: .init(
-                    dnsDuration: .milliseconds(12),
-                    connectDuration: .milliseconds(34),
-                    tlsDuration: .milliseconds(56),
-                    negotiatedTLSVersion: "TLSv1.3",
-                    isProxyConnection: true,
-                    remoteAddress: "192.0.2.10:443"
-                )
-            )
-        ).run()
-
-        #expect(result.evidence.contains(.init(code: "https.metrics.dns-ms", value: "12")))
-        #expect(result.evidence.contains(.init(code: "https.metrics.connect-ms", value: "34")))
-        #expect(result.evidence.contains(.init(code: "https.metrics.tls-ms", value: "56")))
-        #expect(result.evidence.contains(.init(code: "https.metrics.tls-version", value: "TLSv1.3")))
-        #expect(result.evidence.contains(.init(code: "https.metrics.proxy", value: "true")))
-        #expect(!result.evidence.contains { $0.value == "192.0.2.10:443" })
-    }
-
-    @Test("control endpoint success and captive-portal timeout is indeterminate")
-    func controlEndpointMicrosoftTimeoutIsIndeterminate() async {
-        let result = await HTTPSControlEndpointCheck(
-            loader: StubControlLoader(
-                httpsStatus: 200,
-                httpStatus: nil,
-                httpBody: nil,
-                httpErrorCode: String(URLError.timedOut.rawValue)
-            )
-        ).run()
-
-        #expect(result.status == .indeterminate)
-        #expect(result.evidence.contains(.init(code: "https.available", value: "200")))
-        #expect(result.evidence.contains(.init(
-            code: "captive-portal.transport-error",
-            value: String(URLError.timedOut.rawValue)
-        )))
-    }
-
-    @Test("control endpoint success and captive-portal timeout needs attention")
-    func controlEndpointMicrosoftTimeoutConclusion() async {
-        let internet = await HTTPSControlEndpointCheck(
-            loader: StubControlLoader(
-                httpsStatus: 200,
-                httpStatus: nil,
-                httpBody: nil,
-                httpErrorCode: String(URLError.timedOut.rawValue)
-            )
-        ).run()
-
-        let conclusion = NetworkDiagnosticConclusion.evaluate([
-            NetworkDiagnosticResult(id: .path, status: .normal, summary: "path"),
-            NetworkDiagnosticResult(id: .gatewayReachability, status: .normal, summary: "gateway"),
-            NetworkDiagnosticResult(id: .dns, status: .normal, summary: "dns"),
-            internet,
-            NetworkDiagnosticResult(id: .ipv6, status: .skipped, summary: "ipv6"),
-            NetworkDiagnosticResult(id: .proxy, status: .normal, summary: "proxy"),
-        ])
-
-        #expect(conclusion == .needsAttention)
-        #expect(conclusion != .networkUnavailable)
-    }
-
-    @Test("control endpoint failures retain distinct evidence")
-    func controlEndpointFailureEvidence() async {
-        let tlsFailure = await HTTPSControlEndpointCheck(
-            loader: StubControlLoader(
-                httpsStatus: nil,
-                httpsErrorCode: String(URLError.secureConnectionFailed.rawValue),
-                httpStatus: 200,
-                httpBody: "Success"
-            )
-        ).run()
-        let certificateTimeFailure = await HTTPSControlEndpointCheck(
-            loader: StubControlLoader(
-                httpsStatus: nil,
-                httpsErrorCode: String(URLError.serverCertificateNotYetValid.rawValue),
-                httpStatus: 200,
-                httpBody: "Success"
-            )
-        ).run()
-        let certificateValidationFailure = await HTTPSControlEndpointCheck(
-            loader: StubControlLoader(
-                httpsStatus: nil,
-                httpsErrorCode: String(URLError.serverCertificateUntrusted.rawValue),
-                httpStatus: 200,
-                httpBody: "Success"
-            )
-        ).run()
-        let connectivityFailure = await HTTPSControlEndpointCheck(
-            loader: StubControlLoader(
-                httpsStatus: nil,
-                httpsErrorCode: String(URLError.timedOut.rawValue),
-                httpStatus: 200,
-                httpBody: "Success"
-            )
-        ).run()
-        let httpsStatusFailure = await HTTPSControlEndpointCheck(
-            loader: StubControlLoader(
-                httpsStatus: 503,
-                httpStatus: 200,
-                httpBody: "Success"
-            )
-        ).run()
-        let captivePortalRedirect = await HTTPSControlEndpointCheck(
-            loader: StubControlLoader(
-                httpsStatus: 200,
-                httpStatus: 302,
-                httpBody: nil
-            )
-        ).run()
-
-        #expect(tlsFailure.evidence.contains(.init(
-            code: "https.tls-error",
-            value: String(URLError.secureConnectionFailed.rawValue)
-        )))
-        #expect(certificateTimeFailure.evidence.contains(.init(
-            code: "https.certificate-time-error",
-            value: String(URLError.serverCertificateNotYetValid.rawValue)
-        )))
-        #expect(certificateValidationFailure.evidence.contains(.init(
-            code: "https.certificate-error",
-            value: String(URLError.serverCertificateUntrusted.rawValue)
-        )))
-        #expect(connectivityFailure.evidence.contains(.init(
-            code: "https.connectivity-error",
-            value: String(URLError.timedOut.rawValue)
-        )))
-        #expect(httpsStatusFailure.evidence.contains(.init(code: "https.http-status", value: "503")))
-        #expect(captivePortalRedirect.evidence.contains(.init(code: "captive-portal.redirect", value: "302")))
-        #expect(tlsFailure.summary == String(
-            localized: "network_diagnostics.internet.secure_connection_failure.summary",
-            comment: "Network self-check TLS or certificate failure summary"
-        ))
-        #expect(certificateValidationFailure.summary == tlsFailure.summary)
-        #expect(certificateTimeFailure.summary == String(
-            localized: "network_diagnostics.internet.certificate_time_failure.summary",
-            comment: "Network self-check certificate time failure summary"
-        ))
-        #expect(connectivityFailure.summary == String(
-            localized: "network_diagnostics.internet.connectivity_failure.summary",
-            comment: "Network self-check connectivity failure summary"
-        ))
-
-        let baseResults = [
-            NetworkDiagnosticResult(id: .path, status: .normal, summary: "path"),
-            NetworkDiagnosticResult(id: .gatewayReachability, status: .normal, summary: "gateway"),
-            NetworkDiagnosticResult(id: .dns, status: .normal, summary: "dns"),
-        ]
-        let trailingResults = [
-            NetworkDiagnosticResult(id: .ipv6, status: .skipped, summary: "ipv6"),
-            NetworkDiagnosticResult(id: .proxy, status: .abnormal, summary: "proxy"),
-        ]
-        #expect(NetworkDiagnosticConclusion.evaluate(
-            baseResults + [tlsFailure] + trailingResults
-        ) == .needsAttention)
-        #expect(NetworkDiagnosticConclusion.evaluate(
-            baseResults + [certificateTimeFailure] + trailingResults
-        ) == .needsAttention)
-        #expect(NetworkDiagnosticConclusion.evaluate(
-            baseResults + [certificateValidationFailure] + trailingResults
-        ) == .needsAttention)
-        #expect(NetworkDiagnosticConclusion.evaluate(
-            baseResults + [connectivityFailure] + trailingResults
-        ) == .needsAttention)
-    }
-
-    @Test("control check loads the approved endpoints with an explicit timeout")
-    func controlEndpointRequests() async {
-        let recorder = ControlEndpointTestRecorder()
-        let check = HTTPSControlEndpointCheck(
-            loader: StubControlLoader(
-                httpsStatus: 200,
-                httpStatus: 200,
-                httpBody: "Success",
-                recorder: recorder
-            )
-        )
-
-        _ = await check.run()
-
-        #expect(Set(await recorder.urls) == Set([
-            "https://www.apple.com/",
-            "https://captive.apple.com/hotspot-detect.html",
-        ]))
-        #expect(await recorder.timeouts == [.seconds(5), .seconds(5)])
-    }
-
-    @Test("system control loader disables caching and connectivity waits")
-    func controlEndpointSessionConfiguration() {
-        let configuration = SystemControlEndpointLoader.configuration(timeout: .seconds(5))
-
-        #expect(configuration.requestCachePolicy == .reloadIgnoringLocalCacheData)
-        #expect(configuration.urlCache == nil)
-        #expect(!configuration.waitsForConnectivity)
-        #expect(configuration.timeoutIntervalForRequest == 5)
-        #expect(configuration.timeoutIntervalForResource == 5)
-    }
-
-    @Test("base endpoint configuration does not inherit explicit system proxies")
-    func baseEndpointDisablesExplicitSystemProxy() {
-        let configuration = SystemControlEndpointLoader.configuration(timeout: .seconds(2))
-
-        #expect(configuration.connectionProxyDictionary?.isEmpty == true)
-        #expect(configuration.proxyConfigurations.isEmpty)
-    }
-
-    @Test("app scopes the ATS HTTP exception to the Microsoft captive-portal/proxy endpoint")
-    func temporaryCaptivePortalATSException() {
-        let ats = Bundle.main.object(forInfoDictionaryKey: "NSAppTransportSecurity") as? [String: Any]
-        let domains = ats?["NSExceptionDomains"] as? [String: Any]
-        let exception = domains?["www.msftconnecttest.com"] as? [String: Any]
-
-        #expect(ats?.count == 1)
-        #expect(ats?["NSAllowsArbitraryLoads"] == nil)
-        #expect(domains?.count == 1)
-        #expect(exception?.count == 1)
-        #expect(exception?["NSExceptionAllowsInsecureHTTPLoads"] as? Bool == true)
-        #expect(exception?["NSIncludesSubdomains"] == nil)
-    }
-
-    @Test("production checks run path DNS HTTPS and proxy in dependency order")
-    @MainActor
-    func productionCheckOrder() {
-        #expect(NetworkDiagnosticsViewModel().checkIDs == [.path, .gatewayReachability, .dns, .internet, .proxy, .ipv6])
-    }
-
-    @Test("all successful DNS samples are available")
-    func dnsAllSamplesResolved() async {
-        let result = await DNSResolutionCheck(resolver: StubDNSResolver(.resolved)).run()
-
-        #expect(result.status == .normal)
-        #expect(result.evidence.contains(.init(code: "dns.success-count", value: "3/3")))
-        #expect(result.evidence.contains(.init(code: "dns.failure-count", value: "0/3")))
-        #expect(result.evidence.contains(.init(code: "dns.indeterminate-count", value: "0/3")))
-    }
-
-    @Test("DNS uses the three authoritative third-party probe targets")
-    func dnsProbeTargets() {
-        let check = DNSResolutionCheck(resolver: StubDNSResolver(.resolved))
-
-        #expect(check.probeTargets == [
-            "www.apple.com",
-            "www.microsoft.com",
-            "www.msftconnecttest.com",
-        ])
-    }
-
-    @Test("independent DNS samples start concurrently and aggregate in target order")
-    func dnsSamplesRunConcurrently() async {
-        let resolver = ConcurrentDNSResolver()
-        let result = await DNSResolutionCheck(
-            resolver: resolver,
-            timeout: .seconds(1)
-        ).run()
-
-        #expect(await resolver.maximumInFlight == 3)
-        #expect(result.evidence.contains(.init(code: "dns.success-count", value: "3/3")))
-    }
-
-    @Test("one failed lookup among successful samples is unstable, not unavailable")
-    func dnsQuorum() async {
-        let check = DNSResolutionCheck(
-            resolver: StubDNSResolver(outcomes: [.resolved, .failed, .resolved])
-        )
-
-        let result = await check.run()
-
-        #expect(result.status == .indeterminate)
-        #expect(result.evidence.contains(.init(code: "dns.success-count", value: "2/3")))
-        #expect(result.evidence.contains(.init(code: "dns.failure-count", value: "1/3")))
-    }
-
-    @Test("DNS evidence preserves each sample outcome without exposing host names")
-    func dnsEvidencePreservesSampleOutcomes() async {
-        let result = await DNSResolutionCheck(
-            resolver: MappingDNSResolver(outcomes: [
-                "www.apple.com": .resolved,
-                "www.microsoft.com": .failed,
-                "www.msftconnecttest.com": .indeterminate,
-            ])
-        ).run()
-
-        #expect(result.evidence.contains(.init(code: "dns.sample.apple", value: "resolved")))
-        #expect(result.evidence.contains(.init(code: "dns.sample.microsoft", value: "failed")))
-        #expect(result.evidence.contains(.init(code: "dns.sample.msft-connect-test", value: "indeterminate")))
-        #expect(!result.evidence.contains { $0.value == "www.apple.com" })
-    }
-
-    @Test("all failed DNS samples are unavailable")
-    func dnsAllSamplesFailed() async {
-        let result = await DNSResolutionCheck(resolver: StubDNSResolver(.failed)).run()
-
-        #expect(result.status == .abnormal)
-        #expect(result.evidence.contains(.init(code: "dns.success-count", value: "0/3")))
-        #expect(result.evidence.contains(.init(code: "dns.failure-count", value: "3/3")))
-    }
-
-    @Test("an indeterminate DNS sample is retried once")
-    func dnsIndeterminateSampleRetriesOnce() async {
-        let resolver = StubDNSResolver(outcomes: [.indeterminate, .resolved, .resolved, .resolved])
-        let result = await DNSResolutionCheck(resolver: resolver).run()
-
-        #expect(result.status == .normal)
-        #expect(await resolver.invocationCount == 4)
-    }
-
-    @Test("cancelling DNS sampling does not retry or exceed one attempt per host")
-    func dnsCancellationStopsSampling() async {
-        let resolver = CancellationAwareDNSResolver()
-        let check = DNSResolutionCheck(resolver: resolver, timeout: .seconds(30))
-        let task = Task { await check.run() }
-
-        await resolver.waitForInvocation()
-        task.cancel()
-        _ = await task.value
-
-        #expect(await resolver.invocationCount <= 3)
-    }
-
-    @Test("all indeterminate DNS samples cannot be determined")
-    func dnsAllSamplesIndeterminate() async {
-        let result = await DNSResolutionCheck(resolver: StubDNSResolver(.indeterminate)).run()
-
-        #expect(result.status == .indeterminate)
-        #expect(result.evidence.contains(.init(code: "dns.indeterminate-count", value: "3/3")))
-    }
-
-    @Test("DNS success result does not expose the test domain")
-    func dnsResultHidesTestDomain() async {
-        let testDomain = "example.com"
-        let result = await DNSResolutionCheck(
-            resolver: StubDNSResolver(.resolved),
-            probeTargets: [testDomain, "example.net", "example.org"]
-        ).run()
-
-        #expect(!result.summary.localizedCaseInsensitiveContains(testDomain))
-    }
-
-    @Test("proxy parser deduplicates HTTP HTTPS and SOCKS endpoints")
-    func proxyParsing() {
-        let configuration = SystemProxyConfiguration(settings: [
-            "HTTPEnable": 1,
-            "HTTPProxy": " Proxy.Example ",
-            "HTTPPort": 8080,
-            "HTTPSEnable": 1,
-            "HTTPSProxy": "proxy.example",
-            "HTTPSPort": 8080,
-            "SOCKSEnable": 1,
-            "SOCKSProxy": "socks.example",
-            "SOCKSPort": 1080,
-        ])
-
-        #expect(configuration.endpoints == [
-            ProxyEndpoint(host: "proxy.example", port: 8080),
-            ProxyEndpoint(host: "socks.example", port: 1080),
-        ])
-        #expect(!configuration.hasInvalidExplicitProxy)
-    }
-
-    @Test("proxy parser reads PAC URL and automatic discovery")
-    func pacParsing() {
-        let configuration = SystemProxyConfiguration(settings: [
-            "ProxyAutoConfigEnable": 1,
-            "ProxyAutoConfigURLString": "https://proxy.example/config.pac",
-            "ProxyAutoDiscoveryEnable": 1,
-        ])
-
-        #expect(configuration.pacEnabled)
-        #expect(configuration.pacURL == "https://proxy.example/config.pac")
-        #expect(configuration.autoDiscoveryEnabled)
-    }
-
-    @Test("PAC result chooses DIRECT without probing a stale explicit endpoint")
-    func pacCanChooseDirect() async {
-        let controlURL = URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        let pacURL = URL(string: "https://proxy.example/config.pac")!
-        let recorder = PACResolutionTestRecorder()
-        let resolver = SystemProxyResolver(
-            pacTimeout: .milliseconds(25),
-            configurationResolver: StubProxyConfigurationResolver(.pac(pacURL)),
-            pacResolver: StubPACResolver(.direct, recorder: recorder)
-        )
-        let resolution = await resolver.resolve(for: controlURL)
-
-        #expect(resolution == ProxyCandidateResolution(candidates: [.direct], evidenceCodes: []))
-        #expect(await recorder.requests == [
-            .init(pacURL: pacURL, targetURL: controlURL, timeout: .milliseconds(25)),
-        ])
-    }
-
-    @Test("static proxy candidates preserve HTTP then DIRECT order")
-    func staticProxyCandidateOrder() {
-        let candidates = SystemProxyResolver.candidates(from: [
-            proxyDictionary(type: kCFProxyTypeHTTP, host: "127.0.0.1", port: 7890),
-            proxyDictionary(type: kCFProxyTypeNone),
-        ])
-
-        #expect(candidates == [
-            .http(.init(host: "127.0.0.1", port: 7890)),
-            .direct,
-        ])
-    }
-
-    @Test("PAC proxy candidates preserve PROXY then DIRECT order")
-    func pacProxyCandidateOrder() {
-        let resolution = SystemPACResolver.resolution(from: [
-            proxyDictionary(type: kCFProxyTypeHTTP, host: "proxy.example", port: 8080),
-            proxyDictionary(type: kCFProxyTypeNone),
-        ])
-
-        #expect(resolution == ProxyCandidateResolution(
-            candidates: [
-                .http(.init(host: "proxy.example", port: 8080)),
-                .direct,
-            ],
-            evidenceCodes: []
-        ))
-    }
-
-    @Test("inline PAC scripts execute at their ordered directive position")
-    func inlinePACScriptPreservesDirectiveOrder() async {
-        let target = URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        let script = "function FindProxyForURL(url, host) { return 'SOCKS proxy.example:1080'; }"
-        let first = EffectiveProxy.http(.init(host: "first.example", port: 8080))
-        let inline = EffectiveProxy.socks(.init(host: "proxy.example", port: 1080))
-        let executor = ControlledPACCallbackExecutor()
-        let resolver = SystemProxyResolver(
-            configurationResolver: StubProxyConfigurationResolver([
-                .http(.init(host: "first.example", port: 8080)),
-                .pacScript(script),
-                .direct,
-            ]),
-            pacResolver: SystemPACResolver(callbackExecutor: executor)
-        )
-        let task = Task { await resolver.resolve(for: target) }
-
-        let request = await executor.nextRequest()
-        #expect(request == .init(source: .script(script), targetURL: target))
-        executor.complete(.success(.init(candidates: [inline], evidenceCodes: [])))
-        let resolution = await task.value
-
-        #expect(resolution == .init(candidates: [first, inline, .direct], evidenceCodes: []))
-    }
-
-    @Test("inline PAC dictionaries preserve their JavaScript source")
-    func inlinePACDictionaryPreservesScript() {
-        let script = "function FindProxyForURL(url, host) { return 'DIRECT'; }"
-        let dictionary: NSDictionary = [
-            kCFProxyTypeKey as String: kCFProxyTypeAutoConfigurationJavaScript,
-            kCFProxyAutoConfigurationJavaScriptKey as String: script,
-        ]
-
-        #expect(SystemProxyResolver.directive(from: dictionary) == .pacScript(script))
-    }
-
-    @Test("PAC callback errors resume with execution-failed evidence")
-    func pacCallbackErrorResumesContinuation() async {
-        let target = URL(string: "https://www.apple.com/")!
-        let pacURL = URL(string: "https://proxy.example/config.pac")!
-        let executor = ControlledPACCallbackExecutor()
-        let resolver = SystemPACResolver(callbackExecutor: executor)
-        let task = Task {
-            await resolver.resolve(pacURL: pacURL, targetURL: target, timeout: .seconds(30))
-        }
-
-        let request = await executor.nextRequest()
-        #expect(request == .init(source: .url(pacURL), targetURL: target))
-        executor.complete(.failure)
-        let resolution = await task.value
-
-        #expect(resolution == .init(candidates: [], evidenceCodes: ["pac-execution-failed"]))
-        #expect(executor.executionWasCancelled)
-    }
-
-    @Test("PAC callback cancellation resumes once with cancellation evidence")
-    func pacCallbackCancellationResumesContinuation() async {
-        let target = URL(string: "https://www.apple.com/")!
-        let pacURL = URL(string: "https://proxy.example/config.pac")!
-        let executor = ControlledPACCallbackExecutor()
-        let resolver = SystemPACResolver(callbackExecutor: executor)
-        let task = Task {
-            await resolver.resolve(pacURL: pacURL, targetURL: target, timeout: .seconds(30))
-        }
-
-        _ = await executor.nextRequest()
-        task.cancel()
-        let resolution = await task.value
-
-        #expect(resolution == .init(candidates: [], evidenceCodes: ["pac-cancelled"]))
-        #expect(executor.executionWasCancelled)
-    }
-
-    @Test("PAC callback timeout resumes once and cancels execution")
-    func pacCallbackTimeoutResumesContinuation() async {
-        let target = URL(string: "https://www.apple.com/")!
-        let pacURL = URL(string: "https://proxy.example/config.pac")!
-        let executor = ControlledPACCallbackExecutor()
-        let resolver = SystemPACResolver(callbackExecutor: executor)
-        let task = Task {
-            await resolver.resolve(pacURL: pacURL, targetURL: target, timeout: .milliseconds(10))
-        }
-
-        _ = await executor.nextRequest()
-        let resolution = await task.value
-
-        #expect(resolution == .init(candidates: [], evidenceCodes: ["pac-timeout"]))
-        #expect(executor.executionWasCancelled)
-    }
-
-    @Test("invalid proxy candidate preserves later valid candidate and evidence")
-    func invalidProxyCandidatePreservesLaterCandidate() async {
-        let validProxy = EffectiveProxy.http(.init(host: "proxy.example", port: 8080))
-        let resolver = SystemProxyResolver(
-            configurationResolver: StubProxyConfigurationResolver(
-                SystemProxyResolver.directives(from: [
-                    proxyDictionary(type: kCFProxyTypeHTTP),
-                    proxyDictionary(
-                        type: kCFProxyTypeHTTP,
-                        host: "proxy.example",
-                        port: 8080
-                    ),
-                ])
-            ),
-            pacResolver: StubPACResolver(.direct)
-        )
-
-        let resolution = await resolver.resolve(
-            for: URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        )
-
-        #expect(resolution.candidates == [validProxy])
-        #expect(resolution.evidenceCodes == ["proxy-endpoint-invalid"])
-    }
-
-    @Test("whitespace-only proxy hosts are invalid resolution entries")
-    func whitespaceOnlyProxyHostIsInvalid() {
-        let directive = SystemProxyResolver.directive(from: proxyDictionary(
-            type: kCFProxyTypeHTTP,
-            host: " \n\t ",
-            port: 8080
-        ))
-
-        #expect(directive == .unavailable("proxy-endpoint-invalid"))
-    }
-
-    @Test("CFNetwork PAC DIRECT result maps to direct egress")
-    func pacDirectParsing() {
-        let result: NSDictionary = [
-            kCFProxyTypeKey as String: kCFProxyTypeNone,
-        ]
-
-        #expect(SystemProxyResolver.directive(from: result) == .direct)
-    }
-
-    @Test("DIRECT does not probe a stale explicit endpoint or proxy egress")
-    func directSkipsStaleProxy() async {
-        let connectorRecorder = ProxyConnectorTestRecorder()
-        let egressRecorder = ProxyEgressTestRecorder()
-        let check = SystemProxyCheck(
-            resolver: StubProxyResolver(.direct),
-            connector: RecordingProxyConnector(reachable: false, recorder: connectorRecorder),
-            egressLoader: StubProxyEgressLoader(statusCode: nil, errorCode: "unexpected", recorder: egressRecorder),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        )
-
-        let result = await check.run()
-
-        #expect(result.status == .indeterminate)
-        #expect(await connectorRecorder.endpoints.isEmpty)
-        #expect(await egressRecorder.requests.isEmpty)
-    }
-
-    @Test("routed tunnel interface is recognized as the proxy route")
-    func tunneledRoutingIsRecognized() async {
-        let check = SystemProxyCheck(
-            resolver: StubProxyResolver(.direct),
-            connector: StubProxyConnector(reachable: false),
-            egressLoader: StubProxyEgressLoader(statusCode: nil, errorCode: "unexpected"),
-            tunnelReader: StubProxyTunnelStateReader(interface: "utun98")
-        )
-
-        let result = await check.run()
-
-        #expect(result.status == .normal)
-        #expect(result.summary == String(
-            localized: "network_diagnostics.proxy.tunnel_routes.summary",
-            comment: "Network self-check tunneled proxy route result summary"
-        ))
-        #expect(result.evidence.contains(.init(code: "proxy.http.route-type", value: "tunnel")))
-        #expect(result.evidence.contains(.init(code: "proxy.https.route-type", value: "tunnel")))
-        #expect(result.evidence.contains(.init(code: "proxy.http.tunnel-interface", value: "utun98")))
-        #expect(result.evidence.contains(.init(code: "proxy.https.tunnel-interface", value: "utun98")))
-        #expect(result.evidence.contains(.init(code: "proxy.http.tunnel-detection-source", value: "nwpath")))
-        #expect(result.evidence.contains(.init(code: "proxy.https.tunnel-detection-source", value: "nwpath")))
-        #expect(result.evidence.contains(.init(code: "proxy.http.egress-status", value: "base-check")))
-    }
-
-    @Test("active interface tunnel detection uses the softer summary")
-    func activeInterfaceTunnelUsesSofterSummary() async {
-        let check = SystemProxyCheck(
-            resolver: StubProxyResolver(.direct),
-            connector: StubProxyConnector(reachable: false),
-            egressLoader: StubProxyEgressLoader(statusCode: nil, errorCode: "unexpected"),
-            tunnelReader: StubProxyTunnelStateReader(
-                interface: "utun98",
-                source: .activeInterface
-            )
-        )
-
-        let result = await check.run()
-
-        #expect(result.status == .normal)
-        #expect(result.summary == String(
-            localized: "network_diagnostics.proxy.tunnel_routes_active.summary",
-            comment: "Network self-check active tunnel interface proxy route result summary"
-        ))
-        #expect(result.evidence.contains(.init(code: "proxy.http.route-type", value: "tunnel")))
-        #expect(result.evidence.contains(.init(code: "proxy.http.tunnel-detection-source", value: "active-interface")))
-        #expect(result.evidence.contains(.init(code: "proxy.https.tunnel-detection-source", value: "active-interface")))
-    }
-
-    @Test("tunnel candidate prefers the routed interface then an active ipv4 tunnel")
-    func tunnelCandidatePriority() {
-        #expect(SystemProxyTunnelStateReader.candidateTunnelState(
-            pathRouted: "utun3",
-            activeIPv4Tunnels: ["utun98"]
-        ) == ProxyTunnelState(interface: "utun3", source: .nwpath))
-        #expect(SystemProxyTunnelStateReader.candidateTunnelState(
-            pathRouted: nil,
-            activeIPv4Tunnels: ["utun98"]
-        ) == ProxyTunnelState(interface: "utun98", source: .activeInterface))
-        #expect(SystemProxyTunnelStateReader.candidateTunnelState(
-            pathRouted: nil,
-            activeIPv4Tunnels: []
-        ) == nil)
-    }
-
-    @Test("tunnel path wait returns nil when the timeout fires before any path")
-    func tunnelPathWaitTimeoutWins() async {
-        let streamTerminated = LockedFlag()
-        let neverEnding = AsyncStream<NWPath> { continuation in
-            continuation.onTermination = { _ in streamTerminated.set() }
-        }
-
-        let path = await SystemProxyTunnelStateReader.firstPath(
-            from: neverEnding,
-            timeout: {}
-        )
-
-        #expect(path == nil)
-        #expect(streamTerminated.isSet)
-    }
-
-    @Test("tunnel path wait cancels the pending timeout when the path side completes first")
-    func tunnelPathWaitCancelsPendingTimeout() async {
-        let probe = TunnelTimeoutCancellationProbe()
-        let finished = AsyncStream<NWPath> { continuation in
-            continuation.finish()
-        }
-
-        let path = await SystemProxyTunnelStateReader.firstPath(
-            from: finished,
-            timeout: { await probe.park() }
-        )
-
-        #expect(path == nil)
-        #expect(probe.didCancel)
-    }
-
-    @Test("failed proxy candidate falls back to DIRECT for the same target")
-    func failedProxyCandidateFallsBackToDirect() async {
-        let httpURL = URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        let httpsURL = URL(string: "https://www.apple.com/")!
-        let staleEndpoint = ProxyEndpoint(host: "stale.example", port: 8080)
-        let connectorRecorder = ProxyConnectorTestRecorder()
-        let egressRecorder = ProxyEgressTestRecorder()
-        let check = SystemProxyCheck(
-            resolver: RecordingProxyResolver(resolutions: [
-                httpURL: .init(candidates: [.http(staleEndpoint), .direct], evidenceCodes: []),
-                httpsURL: .init(candidates: [.direct], evidenceCodes: []),
-            ]),
-            connector: RecordingProxyConnector(reachable: false, recorder: connectorRecorder),
-            egressLoader: StubProxyEgressLoader(
-                statusCode: 200,
-                recorder: egressRecorder
-            ),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        )
-
-        let result = await check.run()
-
-        #expect(result.status == .indeterminate)
-        #expect(result.evidence.contains(.init(code: "proxy.http.endpoint-status", value: "unavailable")))
-        #expect(result.evidence.contains(.init(code: "proxy.http.candidate-index", value: "1")))
-        #expect(result.evidence.contains(.init(code: "proxy.http.route-type", value: "direct")))
-        #expect(result.evidence.contains(.init(code: "proxy.http.fallback-used", value: "true")))
-        #expect(await connectorRecorder.endpoints == [staleEndpoint])
-        #expect(await egressRecorder.requests.isEmpty)
-    }
-
-    @Test("failed first proxy tries the next proxy candidate")
-    func failedFirstProxyTriesNextProxy() async {
-        let httpURL = URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        let httpsURL = URL(string: "https://www.apple.com/")!
-        let staleEndpoint = ProxyEndpoint(host: "stale.example", port: 8080)
-        let workingEndpoint = ProxyEndpoint(host: "working.example", port: 8081)
-        let selectedProxy = EffectiveProxy.http(workingEndpoint)
-        let connector = SequencedProxyConnector(outcomes: [false, true])
-        let egressRecorder = ProxyEgressTestRecorder()
-        let check = SystemProxyCheck(
-            resolver: RecordingProxyResolver(resolutions: [
-                httpURL: .init(
-                    candidates: [.http(staleEndpoint), selectedProxy],
-                    evidenceCodes: []
-                ),
-                httpsURL: .init(candidates: [.direct], evidenceCodes: []),
-            ]),
-            connector: connector,
-            egressLoader: StubProxyEgressLoader(statusCode: 204, recorder: egressRecorder),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        )
-
-        let result = await check.run()
-
-        #expect(result.status == .indeterminate)
-        #expect(await connector.endpoints == [staleEndpoint, workingEndpoint])
-        #expect(await egressRecorder.requests.map(\.proxy) == [selectedProxy])
-        #expect(result.evidence.contains(.init(code: "proxy.http.candidate-index", value: "1")))
-        #expect(result.evidence.contains(.init(code: "proxy.http.fallback-used", value: "true")))
-    }
-
-    @Test("candidate attempts share one controlled overall timeout")
-    func proxyCandidateAttemptsShareTimeout() async {
-        let target = URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        let firstProxy = EffectiveProxy.http(.init(host: "first.example", port: 8080))
-        let secondProxy = EffectiveProxy.http(.init(host: "second.example", port: 8081))
-        let connector = SequencedProxyConnector(outcomes: [true, true])
-        let egressLoader = SequencedProxyEgressLoader(
-            responses: [
-                .init(statusCode: nil, errorCode: "timed-out"),
-                .init(statusCode: 200, errorCode: nil),
-            ],
-            delays: [.zero, .zero]
-        )
-        let clock = SequencedProxyCheckClock(offsets: [
-            .zero,
-            .zero,
-            .milliseconds(250),
-            .milliseconds(1_200),
-            .milliseconds(1_300),
-        ])
-        let check = SystemProxyCheck(
-            resolver: RecordingProxyResolver(resolutions: [:]),
-            connector: connector,
-            egressLoader: egressLoader,
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-            timeout: .seconds(2),
-            clock: clock
-        )
-
-        let result = await check.evaluate(
-            target: target,
-            resolution: .init(candidates: [firstProxy, secondProxy], evidenceCodes: []),
-            timeout: .seconds(2)
-        )
-
-        #expect(result.status == .proxied)
-        #expect(result.selectedCandidateIndex == 1)
-        #expect(result.selectedProxy == secondProxy)
-        #expect(await connector.endpoints == [
-            .init(host: "first.example", port: 8080),
-            .init(host: "second.example", port: 8081),
-        ])
-        #expect(await connector.timeouts == [.seconds(1), .milliseconds(800)])
-        #expect(await egressLoader.timeouts == [.milliseconds(750), .milliseconds(700)])
-    }
-
-    @Test("an expired overall proxy timeout does not start another candidate")
-    func expiredProxyTimeoutStopsCandidateFallback() async {
-        let target = URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        let firstProxy = EffectiveProxy.http(.init(host: "first.example", port: 8080))
-        let secondProxy = EffectiveProxy.http(.init(host: "second.example", port: 8081))
-        let connector = SequencedProxyConnector(outcomes: [true, true])
-        let egressLoader = SequencedProxyEgressLoader(
-            responses: [.init(statusCode: nil, errorCode: "timed-out")],
-            delays: [.zero]
-        )
-        let clock = SequencedProxyCheckClock(offsets: [
-            .zero,
-            .zero,
-            .milliseconds(250),
-            .milliseconds(2_100),
-        ])
-        let check = SystemProxyCheck(
-            resolver: RecordingProxyResolver(resolutions: [:]),
-            connector: connector,
-            egressLoader: egressLoader,
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-            timeout: .seconds(2),
-            clock: clock
-        )
-
-        let result = await check.evaluate(
-            target: target,
-            resolution: .init(candidates: [firstProxy, secondProxy], evidenceCodes: []),
-            timeout: .seconds(2)
-        )
-
-        #expect(result.status == .unavailable)
-        #expect(await connector.endpoints == [.init(host: "first.example", port: 8080)])
-        #expect(await connector.timeouts == [.seconds(1)])
-        #expect(await egressLoader.timeouts == [.milliseconds(750)])
-        #expect(result.evidence.contains(.init(code: "proxy.http.timeout", value: "expired")))
-    }
-
-    @Test("proxy check resolves the HTTP and HTTPS targets independently")
-    func proxyTargetsAreResolvedIndependently() async {
-        let recorder = ProxyResolutionTestRecorder()
-        let httpURL = URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        let httpsURL = URL(string: "https://www.apple.com/")!
-        let httpsProxy = EffectiveProxy.https(.init(host: "secure-proxy.example", port: 8443))
-        let resolver = RecordingProxyResolver(
-            resolutions: [
-                httpURL: .init(candidates: [.direct], evidenceCodes: []),
-                httpsURL: .init(candidates: [httpsProxy], evidenceCodes: []),
-            ],
-            recorder: recorder
-        )
-        let check = SystemProxyCheck(
-            resolver: resolver,
-            connector: StubProxyConnector(reachable: true),
-            egressLoader: StubProxyEgressLoader(statusCode: 200),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        )
-
-        let result = await check.run()
-        let resolvedURLs = await recorder.urls
-
-        #expect(resolvedURLs.count == 2)
-        #expect(Set(resolvedURLs) == Set([httpURL.absoluteString, httpsURL.absoluteString]))
-        #expect(result.status == .indeterminate)
-        #expect(result.summary == String(
-            localized: "network_diagnostics.proxy.mixed_routing.summary",
-            comment: "Network self-check mixed target proxy routing result summary"
-        ))
-    }
-
-    @Test("both tested targets can use direct routes without a global disabled claim")
-    func bothProxyTargetsDirect() async {
-        let recorder = ProxyResolutionTestRecorder()
-        let resolver = RecordingProxyResolver(
-            defaultResolution: .init(candidates: [.direct], evidenceCodes: []),
-            recorder: recorder
-        )
-
-        let result = await SystemProxyCheck(
-            resolver: resolver,
-            connector: StubProxyConnector(reachable: false),
-            egressLoader: StubProxyEgressLoader(statusCode: nil, errorCode: "unexpected"),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        ).run()
-
-        #expect(result.status == .indeterminate)
-        #expect(result.summary == String(
-            localized: "network_diagnostics.proxy.direct_routes.summary",
-            comment: "Network self-check direct target routes result summary"
-        ))
-        #expect(result.summary != String(
-            localized: "network_diagnostics.proxy.disabled.summary",
-            comment: "Network self-check system proxy disabled result summary"
-        ))
-        #expect(Set(await recorder.urls) == Set([
-            "http://www.msftconnecttest.com/connecttest.txt",
-            "https://www.apple.com/",
-        ]))
-    }
-
-    @Test("both tested proxy routes are available")
-    func bothProxyTargetsProxied() async {
-        let httpURL = URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        let httpsURL = URL(string: "https://www.apple.com/")!
-        let resolver = RecordingProxyResolver(resolutions: [
-            httpURL: .init(
-                candidates: [.http(.init(host: "proxy.example", port: 8080))],
-                evidenceCodes: []
-            ),
-            httpsURL: .init(
-                candidates: [.https(.init(host: "proxy.example", port: 8443))],
-                evidenceCodes: []
-            ),
-        ])
-
-        let result = await SystemProxyCheck(
-            resolver: resolver,
-            connector: StubProxyConnector(reachable: true),
-            egressLoader: StubProxyEgressLoader(statusCode: 200),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        ).run()
-
-        #expect(result.status == .normal)
-        #expect(result.proxyFacts == .init(http: .available, https: .available))
-        #expect(result.summary == String(
-            localized: "network_diagnostics.proxy.routes_available.summary",
-            comment: "Network self-check proxy target routes available result summary"
-        ))
-    }
-
-    @Test("a tested target reports proxy authentication required")
-    func proxyTargetAuthenticationRequired() async {
-        let httpURL = URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        let httpsURL = URL(string: "https://www.apple.com/")!
-        let resolver = RecordingProxyResolver(resolutions: [
-            httpURL: .init(
-                candidates: [.http(.init(host: "proxy.example", port: 8080))],
-                evidenceCodes: []
-            ),
-            httpsURL: .init(candidates: [.direct], evidenceCodes: []),
-        ])
-
-        let result = await SystemProxyCheck(
-            resolver: resolver,
-            connector: StubProxyConnector(reachable: true),
-            egressLoader: StubProxyEgressLoader(statusCode: 407),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        ).run()
-
-        #expect(result.status == .abnormal)
-        #expect(result.summary == String(
-            localized: "network_diagnostics.proxy.authentication_required.summary",
-            comment: "Network self-check proxy authentication required result summary"
-        ))
-        #expect(result.evidence.contains(.init(
-            code: "proxy.http.authentication-status",
-            value: "required"
-        )))
-        #expect(result.evidence.contains(.init(code: "proxy.http.egress-status", value: "407")))
-    }
-
-    @Test("configured candidates with no working route are unavailable")
-    func proxyTargetRouteUnavailable() async {
-        let resolver = RecordingProxyResolver(defaultResolution: .init(
-            candidates: [.http(.init(host: "stale.example", port: 8080))],
-            evidenceCodes: []
-        ))
-
-        let result = await SystemProxyCheck(
-            resolver: resolver,
-            connector: StubProxyConnector(reachable: false),
-            egressLoader: StubProxyEgressLoader(statusCode: 200),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        ).run()
-
-        #expect(result.status == .abnormal)
-        #expect(result.summary == String(
-            localized: "network_diagnostics.proxy.route_unavailable.summary",
-            comment: "Network self-check proxy target route unavailable result summary"
-        ))
-        #expect(result.evidence.contains(.init(
-            code: "proxy.http.endpoint-status",
-            value: "unavailable"
-        )))
-        #expect(result.evidence.contains(.init(
-            code: "proxy.https.endpoint-status",
-            value: "unavailable"
-        )))
-    }
-
-    @Test("empty target resolution is indeterminate")
-    func emptyProxyTargetResolutionIsIndeterminate() async {
-        let result = await SystemProxyCheck(
-            resolver: RecordingProxyResolver(defaultResolution: .init(
-                candidates: [],
-                evidenceCodes: ["resolution-empty"]
-            )),
-            connector: StubProxyConnector(reachable: true),
-            egressLoader: StubProxyEgressLoader(statusCode: 200),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        ).run()
-
-        #expect(result.status == .indeterminate)
-        #expect(result.summary == String(
-            localized: "network_diagnostics.proxy.unable_to_determine.summary",
-            comment: "Network self-check target proxy routing indeterminate result summary"
-        ))
-        #expect(result.evidence.contains(.init(
-            code: "proxy.http.resolution",
-            value: "resolution-empty"
-        )))
-        #expect(result.evidence.contains(.init(
-            code: "proxy.https.resolution",
-            value: "resolution-empty"
-        )))
-    }
-
-    @Test("PAC timeout is reported as PAC unavailable")
-    func pacTimeout() async {
-        let pacURL = URL(string: "https://proxy.example/config.pac")!
-        let check = SystemProxyCheck(
-            resolver: SystemProxyResolver(
-                configurationResolver: StubProxyConfigurationResolver(.pac(pacURL)),
-                pacResolver: StubPACResolver(.unavailable("pac-timeout"))
-            ),
-            connector: StubProxyConnector(reachable: true),
-            egressLoader: StubProxyEgressLoader(statusCode: 200),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        )
-
-        let result = await check.run()
-
-        #expect(result.status == .indeterminate)
-        #expect(result.summary == String(
-            localized: "network_diagnostics.proxy.unable_to_determine.summary",
-            comment: "Network self-check target proxy routing indeterminate result summary"
-        ))
-        #expect(result.evidence.contains(.init(code: "proxy.http.resolution", value: "pac-timeout")))
-        #expect(result.evidence.contains(.init(code: "proxy.https.resolution", value: "pac-timeout")))
-    }
-
-    @Test("selected proxy endpoint failure is distinct from egress failure")
-    func selectedProxyEndpointUnavailable() async {
-        let egressRecorder = ProxyEgressTestRecorder()
-        let proxy = EffectiveProxy.http(ProxyEndpoint(host: "proxy.example", port: 8080))
-        let check = SystemProxyCheck(
-            resolver: StubProxyResolver(proxy),
-            connector: StubProxyConnector(reachable: false),
-            egressLoader: StubProxyEgressLoader(statusCode: 200, recorder: egressRecorder),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        )
-
-        let result = await check.run()
-
-        #expect(result.status == .abnormal)
-        #expect(result.summary == String(
-            localized: "network_diagnostics.proxy.route_unavailable.summary",
-            comment: "Network self-check proxy target route unavailable result summary"
-        ))
-        #expect(result.evidence.contains(.init(code: "proxy.endpoint-unavailable", value: nil)))
-        #expect(await egressRecorder.requests.isEmpty)
-    }
-
-    @Test("HTTP 407 is proxy authentication required")
-    func proxyAuthenticationRequired() async {
-        let proxy = EffectiveProxy.http(ProxyEndpoint(host: "proxy.example", port: 8080))
-        let check = SystemProxyCheck(
-            resolver: StubProxyResolver(proxy),
-            connector: StubProxyConnector(reachable: true),
-            egressLoader: StubProxyEgressLoader(statusCode: 407),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        )
-
-        let result = await check.run()
-
-        #expect(result.status == .abnormal)
-        #expect(result.summary == String(
-            localized: "network_diagnostics.proxy.authentication_required.summary",
-            comment: "Network self-check proxy authentication required result summary"
-        ))
-        #expect(result.evidence.contains(.init(code: "proxy.authentication-required", value: "407")))
-        #expect(!result.evidence.contains(where: { $0.code == "proxy.endpoint-unavailable" }))
-    }
-
-    @Test("URL loading authentication challenge remains a proxy 407 result")
-    func proxyAuthenticationChallenge() {
-        let response = SystemProxyEgressLoader.response(
-            for: URLError(.userAuthenticationRequired)
-        )
-
-        #expect(response == ProxyEgressResponse(statusCode: 407, errorCode: nil))
-    }
-
-    @Test("production tunneled egress configuration disables selected proxy failover")
-    func selectedProxyDoesNotFailOver() throws {
-        let proxy = EffectiveProxy.https(ProxyEndpoint(host: "proxy.example", port: 8080))
-        let configuration = try #require(
-            SystemProxyEgressLoader.configuration(for: proxy, timeout: .seconds(3))
-        )
-
-        #expect(configuration.proxyConfigurations.count == 1)
-        #expect(configuration.proxyConfigurations[0].allowFailover == false)
-        #expect(configuration.connectionProxyDictionary?.isEmpty != false)
-        #expect(configuration.urlCredentialStorage == nil)
-    }
-
-    @Test("HTTP forward HTTPS tunnel and SOCKS candidates use distinct transport configurations")
-    func selectedProxyTransportConfigurationsPreserveSemantics() throws {
-        let endpoint = ProxyEndpoint(host: "proxy.example", port: 8080)
-        let httpConfiguration = try #require(SystemProxyEgressLoader.configuration(
-            for: .http(endpoint),
-            timeout: .seconds(3)
-        ))
-        let httpsConfiguration = try #require(SystemProxyEgressLoader.configuration(
-            for: .https(endpoint),
-            timeout: .seconds(3)
-        ))
-        let socksConfiguration = try #require(SystemProxyEgressLoader.configuration(
-            for: .socks(endpoint),
-            timeout: .seconds(3)
-        ))
-
-        let httpDictionary = try #require(httpConfiguration.connectionProxyDictionary)
-        #expect((httpDictionary[kCFNetworkProxiesHTTPEnable as String] as? NSNumber)?.boolValue == true)
-        #expect(httpDictionary[kCFNetworkProxiesHTTPProxy as String] as? String == endpoint.host)
-        #expect((httpDictionary[kCFNetworkProxiesHTTPPort as String] as? NSNumber)?.intValue == Int(endpoint.port))
-        #expect((httpDictionary[kCFNetworkProxiesHTTPSEnable as String] as? NSNumber)?.boolValue == false)
-        #expect((httpDictionary[kCFNetworkProxiesSOCKSEnable as String] as? NSNumber)?.boolValue == false)
-        #expect((httpDictionary[kCFNetworkProxiesProxyAutoConfigEnable as String] as? NSNumber)?.boolValue == false)
-        #expect((httpDictionary[kCFNetworkProxiesProxyAutoDiscoveryEnable as String] as? NSNumber)?.boolValue == false)
-        #expect(httpConfiguration.proxyConfigurations.isEmpty)
-
-        #expect(httpsConfiguration.connectionProxyDictionary?.isEmpty == true)
-        #expect(httpsConfiguration.proxyConfigurations.count == 1)
-        #expect(httpsConfiguration.proxyConfigurations[0].debugDescription.contains("http_connect"))
-        #expect(httpsConfiguration.proxyConfigurations[0].allowFailover == false)
-
-        #expect(socksConfiguration.connectionProxyDictionary?.isEmpty == true)
-        #expect(socksConfiguration.proxyConfigurations.count == 1)
-        #expect(socksConfiguration.proxyConfigurations[0].debugDescription.contains("socksv5"))
-        #expect(socksConfiguration.proxyConfigurations[0].allowFailover == false)
-    }
-
-    @Test("non-positive proxy timeout without an attempted endpoint is indeterminate")
-    func nonPositiveProxyTimeoutIsIndeterminate() async {
-        let target = URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        let endpoint = ProxyEndpoint(host: "proxy.example", port: 8080)
-        let connectorRecorder = ProxyConnectorTestRecorder()
-        let check = SystemProxyCheck(
-            resolver: StubProxyResolver(.direct),
-            connector: RecordingProxyConnector(reachable: true, recorder: connectorRecorder),
-            egressLoader: StubProxyEgressLoader(statusCode: 200),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        )
-
-        let route = await check.evaluate(
-            target: target,
-            resolution: .init(candidates: [.http(endpoint)], evidenceCodes: []),
-            timeout: .zero
-        )
-
-        #expect(route.status == .indeterminate)
-        #expect(route.selectedCandidateIndex == nil)
-        #expect(route.selectedProxy == nil)
-        #expect(route.evidence.contains(.init(code: "proxy.http.timeout", value: "expired")))
-        #expect(await connectorRecorder.endpoints.isEmpty)
-    }
-
-    @Test("later success after 407 selects that candidate and skips the trailing route")
-    func proxySuccessAfterAuthenticationShortCircuitsTrailingCandidate() async {
-        let target = URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        let firstEndpoint = ProxyEndpoint(host: "auth.example", port: 8080)
-        let selectedEndpoint = ProxyEndpoint(host: "working.example", port: 8081)
-        let trailingEndpoint = ProxyEndpoint(host: "unused.example", port: 8082)
-        let first = EffectiveProxy.http(firstEndpoint)
-        let selected = EffectiveProxy.http(selectedEndpoint)
-        let trailing = EffectiveProxy.http(trailingEndpoint)
-        let connector = SequencedProxyConnector(outcomes: [true, true, true])
-        let egress = SequencedProxyEgressLoader(
-            responses: [
-                .init(statusCode: 407, errorCode: nil),
-                .init(statusCode: 204, errorCode: nil),
-                .init(statusCode: 200, errorCode: nil),
-            ],
-            delays: [.zero, .zero, .zero]
-        )
-        let check = SystemProxyCheck(
-            resolver: StubProxyResolver(.direct),
-            connector: connector,
-            egressLoader: egress,
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        )
-
-        let route = await check.evaluate(
-            target: target,
-            resolution: .init(candidates: [first, selected, trailing], evidenceCodes: []),
-            timeout: .seconds(3)
-        )
-
-        #expect(route.status == .proxied)
-        #expect(route.selectedCandidateIndex == 1)
-        #expect(route.selectedProxy == selected)
-        #expect(await connector.endpoints == [firstEndpoint, selectedEndpoint])
-        #expect(await egress.timeouts.count == 2)
-    }
-
-    @Test("authentication evidence takes precedence over an unavailable peer target")
-    func proxyMixedFailureSeverityPrefersAuthentication() async {
-        let httpURL = URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        let httpsURL = URL(string: "https://www.apple.com/")!
-        let httpProxy = EffectiveProxy.http(.init(host: "auth.example", port: 8080))
-        let httpsProxy = EffectiveProxy.https(.init(host: "stale.example", port: 8443))
-        let connector = EndpointSelectiveProxyConnector(reachableHosts: ["auth.example"])
-        let result = await SystemProxyCheck(
-            resolver: RecordingProxyResolver(resolutions: [
-                httpURL: .init(candidates: [httpProxy], evidenceCodes: []),
-                httpsURL: .init(candidates: [httpsProxy], evidenceCodes: []),
-            ]),
-            connector: connector,
-            egressLoader: StubProxyEgressLoader(statusCode: 407),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        ).run()
-
-        #expect(result.status == .abnormal)
-        #expect(result.summary == String(
-            localized: "network_diagnostics.proxy.authentication_required.summary",
-            comment: "Network self-check proxy authentication required result summary"
-        ))
-        #expect(result.evidence.contains(.init(code: "proxy.http.authentication-status", value: "required")))
-        #expect(result.evidence.contains(.init(code: "proxy.https.endpoint-status", value: "unavailable")))
-    }
-
-    @Test("selected proxy transport failure is egress unavailable")
-    func proxyEgressUnavailable() async {
-        let proxy = EffectiveProxy.http(ProxyEndpoint(host: "proxy.example", port: 8080))
-        let check = SystemProxyCheck(
-            resolver: StubProxyResolver(proxy),
-            connector: StubProxyConnector(reachable: true),
-            egressLoader: StubProxyEgressLoader(statusCode: nil, errorCode: "-1005"),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        )
-
-        let result = await check.run()
-
-        #expect(result.status == .abnormal)
-        #expect(result.summary == String(
-            localized: "network_diagnostics.proxy.route_unavailable.summary",
-            comment: "Network self-check proxy target route unavailable result summary"
-        ))
-        #expect(result.evidence.contains(.init(code: "proxy.egress-unavailable", value: "-1005")))
-    }
-
-    @Test("selected HTTP proxy probes the same control URL once")
-    func selectedProxyEgressSuccess() async {
-        let controlURL = URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        let httpsURL = URL(string: "https://www.apple.com/")!
-        let proxy = EffectiveProxy.http(ProxyEndpoint(host: "proxy.example", port: 8080))
-        let recorder = ProxyEgressTestRecorder()
-        let check = SystemProxyCheck(
-            resolver: RecordingProxyResolver(resolutions: [
-                controlURL: .init(candidates: [proxy], evidenceCodes: []),
-                httpsURL: .init(candidates: [.direct], evidenceCodes: []),
-            ]),
-            connector: StubProxyConnector(reachable: true),
-            egressLoader: StubProxyEgressLoader(statusCode: 200, recorder: recorder),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        )
-
-        let result = await check.run()
-
-        #expect(result.status == .indeterminate)
-        #expect(await recorder.requests == [.init(url: controlURL, proxy: proxy)])
-    }
-
-    @Test("cancelled concurrent proxy resolution remains bounded")
-    func proxyCancellationStopsSecondTargetResolution() async {
-        let resolver = CancellationIgnoringProxyResolver()
-        let check = SystemProxyCheck(
-            resolver: resolver,
-            connector: StubProxyConnector(reachable: true),
-            egressLoader: StubProxyEgressLoader(statusCode: 200),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        )
-        let task = Task { await check.run() }
-
-        await resolver.waitForInvocation()
-        task.cancel()
-        let result = await task.value
-
-        #expect(await resolver.urls.count <= 2)
-        #expect(result.status == .indeterminate)
-        #expect(result.evidence.contains(.init(
-            code: "proxy.cancelled",
-            value: "after-http-resolution"
-        )))
-    }
-
-    @Test("cancelled PAC resolution does not append later static candidates")
-    func proxyResolverCancellationStopsRemainingDirectives() async {
-        let pacURL = URL(string: "https://proxy.example/config.pac")!
-        let pacResolver = CancellationIgnoringPACResolver()
-        let resolver = SystemProxyResolver(
-            configurationResolver: StubProxyConfigurationResolver([
-                .pac(pacURL),
-                .http(.init(host: "unused.example", port: 8080)),
-            ]),
-            pacResolver: pacResolver
-        )
-        let task = Task {
-            await resolver.resolve(
-                for: URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-            )
-        }
-
-        await pacResolver.waitForInvocation()
-        task.cancel()
-        let resolution = await task.value
-
-        #expect(resolution.candidates.isEmpty)
-        #expect(resolution.evidenceCodes.contains("pac-cancelled"))
-        #expect(resolution.evidenceCodes.contains("resolution-cancelled"))
-    }
-
-    @Test("cancelling an endpoint attempt stops candidate fallback")
-    func proxyCancellationStopsAfterConnector() async {
-        let target = URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        let first = ProxyEndpoint(host: "first.example", port: 8080)
-        let second = ProxyEndpoint(host: "second.example", port: 8081)
-        let connector = CancellationIgnoringProxyConnector()
-        let check = SystemProxyCheck(
-            resolver: StubProxyResolver(.direct),
-            connector: connector,
-            egressLoader: StubProxyEgressLoader(statusCode: 200),
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        )
-        let task = Task {
-            await check.evaluate(
-                target: target,
-                resolution: .init(candidates: [.http(first), .http(second)], evidenceCodes: []),
-                timeout: .seconds(30)
-            )
-        }
-
-        await connector.waitForInvocation()
-        task.cancel()
-        let route = await task.value
-
-        #expect(await connector.endpoints == [first])
-        #expect(route.status == .indeterminate)
-        #expect(route.evidence.contains(.init(
-            code: "proxy.http.cancelled",
-            value: "endpoint"
-        )))
-    }
-
-    @Test("cancelling proxy egress stops connector and candidate fallback")
-    func proxyCancellationStopsAfterEgress() async {
-        let target = URL(string: "http://www.msftconnecttest.com/connecttest.txt")!
-        let first = ProxyEndpoint(host: "first.example", port: 8080)
-        let second = ProxyEndpoint(host: "second.example", port: 8081)
-        let connector = SequencedProxyConnector(outcomes: [true, true])
-        let egress = CancellationIgnoringProxyEgressLoader()
-        let check = SystemProxyCheck(
-            resolver: StubProxyResolver(.direct),
-            connector: connector,
-            egressLoader: egress,
-            tunnelReader: StubProxyTunnelStateReader(interface: nil),
-        )
-        let task = Task {
-            await check.evaluate(
-                target: target,
-                resolution: .init(candidates: [.http(first), .http(second)], evidenceCodes: []),
-                timeout: .seconds(30)
-            )
-        }
-
-        await egress.waitForInvocation()
-        task.cancel()
-        let route = await task.value
-
-        #expect(await connector.endpoints == [first])
-        #expect(await egress.invocationCount == 1)
-        #expect(route.status == .indeterminate)
-        #expect(route.evidence.contains(.init(
-            code: "proxy.http.cancelled",
-            value: "egress"
-        )))
-    }
-
-    @Test("view model starts idle and only runs after start")
-    @MainActor
-    func viewModelManualStart() async {
-        let recorder = DiagnosticTestRecorder()
-        let guidance = IsolatedGuidance()
-        let viewModel = NetworkDiagnosticsViewModel(
-            checks: makeStubChecks(recorder: recorder),
-            fingerprintMonitor: DisabledNetworkFingerprintMonitor(),
-            guidance: guidance.coordinator
-        )
-
-        #expect(viewModel.phase == .idle)
-        #expect(viewModel.conclusion == nil)
-        #expect(await recorder.values.isEmpty)
-
-        #expect(viewModel.start())
-        #expect(!viewModel.start())
-        await viewModel.waitForCompletion()
-
-        #expect(viewModel.phase == .completed)
-        #expect(viewModel.conclusion == .networkNormal)
-        #expect(viewModel.results.count == 3)
-        #expect(viewModel.executionPhases.values.allSatisfy { $0 == .completed })
-    }
-
-    @Test("fingerprint observation timeout does not wait for a cancellation-ignoring monitor")
-    @MainActor
-    func fingerprintObservationTimeoutIsBounded() async {
-        let monitor = CancellationIgnoringNetworkFingerprintMonitor()
-        let clock = ManualDiagnosticClock()
-        let viewModel = NetworkDiagnosticsViewModel(
-            checks: [],
-            fingerprintMonitor: monitor,
-            clock: clock
-        )
-
-        #expect(viewModel.start())
-        await monitor.waitForInvocation()
-        await clock.set(clock.origin.advanced(by: .seconds(1)))
-
-        for _ in 0..<100 {
-            if !viewModel.fingerprintMonitoringAvailable { break }
-            await Task.yield()
-        }
-
-        #expect(!viewModel.fingerprintMonitoringAvailable)
-        await monitor.release()
-        await viewModel.waitForCompletion()
-    }
-
-    @Test("network diagnostics keeps an isolated resettable log buffer")
-    func networkDiagnosticsLogStoreIsResettable() {
-        var store = NetworkDiagnosticsLogStore()
-
-        store.append("self-check started")
-        store.append("DNS resolved")
-
-        #expect(store.lines == ["self-check started", "DNS resolved"])
-        #expect(store.text == "self-check started\nDNS resolved")
-
-        store.reset()
-
-        #expect(store.lines.isEmpty)
-        #expect(store.text.isEmpty)
-    }
-
-    @Test("diagnostics log store keeps a bounded history and truncation marker")
-    func diagnosticsLogStoreHasFiniteCapacity() {
-        var store = NetworkDiagnosticsLogStore()
-
-        for index in 0..<501 {
-            store.append("event \(index)")
-        }
-
-        #expect(store.lines.count == NetworkDiagnosticsLogStore.capacity)
-        #expect(store.lines.contains(NetworkDiagnosticsLogStore.truncationMarker))
-        #expect(!store.text.contains("event 0\n"))
-        #expect(store.text.contains("event 500"))
-
-        var typedStore = NetworkDiagnosticsLogStore()
-        for index in 0..<501 {
-            typedStore.append(NetworkDiagnosticEvent(
-                runID: UUID(),
-                elapsedMilliseconds: Int64(index),
-                kind: .checkFinished,
-                checkID: .path,
-                reasonCode: nil
-            ))
-        }
-        #expect(typedStore.lines.count == NetworkDiagnosticsLogStore.capacity)
-        #expect(typedStore.events.count == NetworkDiagnosticsLogStore.capacity - 1)
-    }
-
-    @Test("typed diagnostic events only format allowlisted restart reasons")
-    func diagnosticEventFormattingAllowlist() {
-        let event = NetworkDiagnosticEvent(
-            runID: UUID(),
-            elapsedMilliseconds: 42,
-            kind: .restarted,
-            checkID: nil,
-            reasonCode: "raw-sensitive-value"
-        )
-
-        #expect(event.formatted() == "42ms · Network changed; restarting (network state)")
-        #expect(!event.formatted().contains("raw-sensitive-value"))
-    }
-
-    @Test("view model publishes its own diagnostic result logs")
-    @MainActor
-    func viewModelPublishesDiagnosticLogs() async {
-        let viewModel = NetworkDiagnosticsViewModel(
-            checks: makeStubChecks(recorder: DiagnosticTestRecorder()),
-            fingerprintMonitor: DisabledNetworkFingerprintMonitor()
-        )
-
-        #expect(viewModel.logText.isEmpty)
-        #expect(viewModel.start())
-        await viewModel.waitForCompletion()
-
-        let stableLogLines = viewModel.logStore.lines.map { line in
-            guard let separator = line.range(of: " · ") else { return line }
-            return String(line[separator.upperBound...])
-        }
-        #expect(stableLogLines == [
-            "Session started",
-            "Run started",
-            "Checking Network Path…",
-            "Network Path finished",
-            "Checking DNS Resolution…",
-            "DNS Resolution finished",
-            "Checking System Proxy…",
-            "System Proxy finished",
-            "Check completed",
-        ])
-        #expect(viewModel.logStore.events.map(\.kind) == [
-            .sessionStarted,
-            .runStarted,
-            .checkStarted,
-            .checkFinished,
-            .checkStarted,
-            .checkFinished,
-            .checkStarted,
-            .checkFinished,
-            .completed,
-        ])
-
-        viewModel.clearLogs()
-        #expect(viewModel.logText.isEmpty)
-    }
-
-    @Test("view model clears the previous conclusion before a rerun")
-    @MainActor
-    func viewModelRerun() async {
-        let recorder = DiagnosticTestRecorder()
-        let guidance = IsolatedGuidance()
-        let viewModel = NetworkDiagnosticsViewModel(
-            checks: makeStubChecks(recorder: recorder),
-            fingerprintMonitor: DisabledNetworkFingerprintMonitor(),
-            guidance: guidance.coordinator
-        )
-
-        #expect(viewModel.start())
-        await viewModel.waitForCompletion()
-        #expect(viewModel.conclusion == .networkNormal)
-
-        #expect(viewModel.start())
-        #expect(viewModel.conclusion == nil)
-        await viewModel.waitForCompletion()
-        #expect(await recorder.values.count == 6)
-    }
-
-    @Test("a changed fingerprint restarts network probes once and retains configuration results")
-    @MainActor
-    func fingerprintChangeRestartsOnce() async {
-        let initial = makeFingerprint(interfaceName: "en0", dnsHash: 1)
-        let fingerprintMonitor = ControlledNetworkFingerprintMonitor(initial: initial)
-        let guidance = IsolatedGuidance()
-        let configurationProbe = RestartableDiagnosticProbe()
-        let networkProbe = RestartableDiagnosticProbe(blockFirstInvocation: true)
-        let checks: [any DiagnosticCheck] = [
-            ProbeDiagnosticCheck(
-                id: .proxy,
-                result: .init(id: .proxy, status: .normal, summary: "configuration"),
-                rerunPolicy: .configurationOnly,
-                probe: configurationProbe
-            ),
-            ProbeDiagnosticCheck(
-                id: .path,
-                result: .init(id: .path, status: .normal, summary: "network"),
-                rerunPolicy: .networkSensitive,
-                probe: networkProbe
-            ),
-        ]
-        let viewModel = NetworkDiagnosticsViewModel(
-            checks: checks,
-            fingerprintMonitor: fingerprintMonitor,
-            guidance: guidance.coordinator
-        )
-
-        #expect(viewModel.start())
-        await networkProbe.waitForInvocationCount(1)
-        await fingerprintMonitor.send(makeFingerprint(interfaceName: "en1", dnsHash: 2))
-        await fingerprintMonitor.send(makeFingerprint(interfaceName: "en2", dnsHash: 3))
-        await viewModel.waitForCompletion()
-
-        #expect(viewModel.phase == .completed)
-        #expect(viewModel.conclusion == .networkNormal)
-        #expect(viewModel.automaticRestartCount == 1)
-        #expect(await configurationProbe.invocationCount == 1)
-        #expect(await networkProbe.invocationCount == 2)
-        #expect(viewModel.results[.proxy]?.summary == "configuration")
-        #expect(viewModel.results[.path]?.summary == "network")
-    }
-
-    @Test("an unchanged fingerprint does not restart an active probe")
-    @MainActor
-    func unchangedFingerprintDoesNotRestart() async {
-        let fingerprint = makeFingerprint(interfaceName: "en0", dnsHash: 1)
-        let fingerprintMonitor = ControlledNetworkFingerprintMonitor(initial: fingerprint)
-        let guidance = IsolatedGuidance()
-        let probe = RestartableDiagnosticProbe(blockFirstInvocation: true)
-        let viewModel = NetworkDiagnosticsViewModel(
-            checks: [
-                ProbeDiagnosticCheck(
-                    id: .path,
-                    result: .init(id: .path, status: .normal, summary: "network"),
-                    rerunPolicy: .networkSensitive,
-                    probe: probe
-                ),
-            ],
-            fingerprintMonitor: fingerprintMonitor,
-            guidance: guidance.coordinator
-        )
-
-        #expect(viewModel.start())
-        await probe.waitForInvocationCount(1)
-        await fingerprintMonitor.send(fingerprint)
-        await probe.releaseFirstInvocation()
-        await viewModel.waitForCompletion()
-
-        #expect(viewModel.automaticRestartCount == 0)
-        #expect(await probe.invocationCount == 1)
-    }
-
-    @Test("a second network change during the rerun starts a fresh run")
-    @MainActor
-    func repeatedFingerprintChangesRestartAgain() async {
-        let initial = makeFingerprint(interfaceName: "en0", dnsHash: 1)
-        let fingerprintMonitor = ControlledNetworkFingerprintMonitor(initial: initial)
-        let probe = BlockingDiagnosticProbe()
-        let guidance = IsolatedGuidance()
-        let viewModel = NetworkDiagnosticsViewModel(
-            checks: [BlockingProbeDiagnosticCheck(id: .path, probe: probe)],
-            fingerprintMonitor: fingerprintMonitor,
-            guidance: guidance.coordinator
-        )
-
-        #expect(viewModel.start())
-        await probe.waitForInvocationCount(1)
-        await fingerprintMonitor.send(makeFingerprint(interfaceName: "en1", dnsHash: 2))
-        await probe.waitForInvocationCount(2)
-        await fingerprintMonitor.send(makeFingerprint(interfaceName: "en2", dnsHash: 3))
-        await probe.waitForInvocationCount(3)
-        await probe.release(invocation: 3)
-        await viewModel.waitForCompletion()
-
-        #expect(viewModel.phase == .completed)
-        #expect(viewModel.automaticRestartCount == 2)
-        #expect(await probe.invocationCount == 3)
-    }
-
-    @Test("the controller closes network-change intake before results are published")
-    func fingerprintControllerFinalizationGate() async {
-        let controller = NetworkDiagnosticRestartController()
-        let fingerprint = makeFingerprint(interfaceName: "en1", dnsHash: 2)
-
-        #expect(await controller.completeRun() == false)
-        #expect(await controller.observe(fingerprint) == false)
-    }
-
-    @Test("stability window starts from the latest network change")
-    func stabilityWindowUsesLatestChange() async {
-        let controller = NetworkDiagnosticRestartController()
-        let clock = ManualDiagnosticClock()
-        let base = clock.origin
-        let first = makeFingerprint(interfaceName: "en1", dnsHash: 2)
-        let second = makeFingerprint(interfaceName: "en2", dnsHash: 3)
-        let completion = OptionalBoolRecorder()
-
-        #expect(await controller.observe(first, at: base))
-        let waitTask = Task {
-            let result = await controller.waitForStability(
-                using: clock,
-                until: base.advanced(by: .seconds(2))
-            )
-            await completion.record(result)
-            return result
-        }
-
-        await clock.waitForSleeperCount(1)
-        await clock.set(base.advanced(by: .milliseconds(400)))
-        #expect(await controller.observe(
-            second,
-            at: base.advanced(by: .milliseconds(400))
-        ))
-
-        await clock.waitForSleeperCount(1)
-        await clock.set(base.advanced(by: .milliseconds(800)))
-        await Task.yield()
-        #expect(await completion.value == nil)
-
-        await clock.set(base.advanced(by: .milliseconds(900)))
-        #expect(await waitTask.value)
-        #expect(await completion.value == true)
-    }
-
-    @Test("explicit cancellation is latched before a run is installed")
-    func restartControllerLatchesCancellationBeforeInstall() async {
-        let controller = NetworkDiagnosticRestartController()
-        let run = Task<[NetworkDiagnosticResult], Never> {
-            try? await Task.sleep(for: .seconds(30))
-            return []
-        }
-
-        await controller.cancelCurrentRun()
-        await controller.install(run)
-
-        #expect(run.isCancelled)
-        run.cancel()
-        _ = await run.value
-    }
-
-    @Test("a duplicate path update is ignored")
-    func duplicatePathUpdateIsIgnored() async throws {
-        let baseline = makeFingerprint(interfaceName: "en0", dnsHash: 1)
-        let path = NetworkPathFingerprint(
-            interfaceType: baseline.interfaceType,
-            interfaceName: baseline.interfaceName,
-            pathStatus: baseline.pathStatus
-        )
-        let monitor = SystemNetworkFingerprintMonitor(
-            settingsReader: MutableFingerprintSettingsReader(dnsHash: 1, proxyHash: 7),
-            pathSource: FiniteNetworkPathFingerprintSource(values: [path, path]),
-            settingsPoller: SilentNetworkFingerprintSettingsPoller()
-        )
-
-        let optionalObservation = await monitor.observation()
-        let observation = try #require(optionalObservation)
-        var changes: [NetworkFingerprint] = []
-        for await fingerprint in observation.changes {
-            changes.append(fingerprint)
-        }
-
-        #expect(observation.baseline == baseline)
-        #expect(changes.isEmpty)
-    }
-
-    @Test("baseline and an immediate path change share one buffered observation source")
-    func fingerprintObservationDoesNotLoseBaselineGapChange() async throws {
-        let baselinePath = NetworkPathFingerprint(
-            interfaceType: "wifi",
-            interfaceName: "en0",
-            pathStatus: .satisfied
-        )
-        let changedPath = NetworkPathFingerprint(
-            interfaceType: "wifi",
-            interfaceName: "en1",
-            pathStatus: .satisfied
-        )
-        let pathSource = BufferedGapNetworkPathFingerprintSource(
-            values: [baselinePath, changedPath]
-        )
-        let monitor = SystemNetworkFingerprintMonitor(
-            settingsReader: MutableFingerprintSettingsReader(dnsHash: 1, proxyHash: 7),
-            pathSource: pathSource,
-            settingsPoller: SilentNetworkFingerprintSettingsPoller()
-        )
-
-        let optionalObservation = await monitor.observation()
-        let observation = try #require(optionalObservation)
-        var iterator = observation.changes.makeAsyncIterator()
-        let change = await iterator.next()
-
-        #expect(observation.baseline.interfaceName == "en0")
-        #expect(change?.interfaceName == "en1")
-        #expect(pathSource.streamRequestCount == 1)
-    }
-
-    @Test("explicit cancellation stops an active diagnostics session")
-    @MainActor
-    func viewModelCancellation() async {
-        let probe = BlockingDiagnosticProbe()
-        let viewModel = NetworkDiagnosticsViewModel(
-            checks: [BlockingProbeDiagnosticCheck(id: .path, probe: probe)],
-            fingerprintMonitor: DisabledNetworkFingerprintMonitor()
-        )
-
-        #expect(viewModel.start())
-        await probe.waitForInvocationCount(1)
-        viewModel.cancel()
-        await viewModel.waitForCompletion()
-
-        #expect(viewModel.phase == .completed)
-        #expect(viewModel.conclusion == nil)
-        #expect(viewModel.endReason == .cancelled)
-    }
-
-    @Test("a cancelled run cannot overwrite a replacement run")
-    @MainActor
-    func cancelledRunCannotOverwriteReplacementRun() async {
-        let probe = CancellationIgnoringDiagnosticProbe()
-        let viewModel = NetworkDiagnosticsViewModel(
-            checks: [CancellationIgnoringProbeDiagnosticCheck(probe: probe)],
-            fingerprintMonitor: DisabledNetworkFingerprintMonitor()
-        )
-
-        #expect(viewModel.start())
-        await probe.waitForInvocationCount(1)
-        viewModel.cancel()
-        #expect(viewModel.start())
-        await probe.waitForInvocationCount(2)
-
-        await probe.release(invocation: 1)
-        try? await Task.sleep(for: .milliseconds(10))
-        #expect(viewModel.phase == .running)
-        #expect(viewModel.results.isEmpty)
-
-        await probe.release(invocation: 2)
-        await viewModel.waitForCompletion()
-
-        #expect(viewModel.phase == .completed)
-        #expect(viewModel.endReason == .completed)
-        #expect(viewModel.results[.path]?.status == .normal)
-    }
-
-    @Test("unstable network changes do not restore invalidated results")
-    @MainActor
-    func unstableNetworkChangeDoesNotRestoreOldResults() async {
-        let initial = makeFingerprint(interfaceName: "en0", dnsHash: 1)
-        let fingerprintMonitor = ControlledNetworkFingerprintMonitor(initial: initial)
-        let blockingProbe = BlockingDiagnosticProbe()
-        let clock = ManualDiagnosticClock()
-        let viewModel = NetworkDiagnosticsViewModel(
-            checks: [
-                StubDiagnosticCheck(
-                    id: .path,
-                    result: .init(id: .path, status: .normal, summary: "old path"),
-                    recorder: DiagnosticTestRecorder()
-                ),
-                BlockingProbeDiagnosticCheck(id: .dns, probe: blockingProbe),
-            ],
-            fingerprintMonitor: fingerprintMonitor,
-            clock: clock
-        )
-
-        #expect(viewModel.start())
-        await blockingProbe.waitForInvocationCount(1)
-        await clock.set(clock.origin.advanced(by: .milliseconds(29_800)))
-        await fingerprintMonitor.send(makeFingerprint(interfaceName: "en1", dnsHash: 1))
-        await clock.set(clock.origin.advanced(by: .milliseconds(30_100)))
-        await viewModel.waitForCompletion()
-
-        #expect(viewModel.phase == .completed)
-        #expect(viewModel.endReason == .superseded)
-        #expect(viewModel.results.isEmpty)
-        #expect(viewModel.pendingCheckIDs == [.path, .dns])
-    }
-
-    @Test("a successful diagnostics run reports the completion moment exactly once")
-    @MainActor
-    func successfulRunReportsDiagnosticsMomentOnce() async {
-        let recorder = DiagnosticTestRecorder()
-        let guidance = IsolatedGuidance()
-        let viewModel = NetworkDiagnosticsViewModel(
-            checks: makeStubChecks(recorder: recorder),
-            fingerprintMonitor: DisabledNetworkFingerprintMonitor(),
-            guidance: guidance.coordinator
-        )
-
-        #expect(viewModel.start())
-        await viewModel.waitForCompletion()
-
-        #expect(viewModel.phase == .completed)
-        let loaded = guidance.store.load()
-        #expect(loaded.meaningfulCompletionCount == 1)
-        #expect(loaded.invitationPresentationCount == 0)
-        #expect(loaded.lastInvitationDate == nil)
-        #expect(guidance.events.filter { $0.name == "guidance.value_moment" }.count == 1)
-    }
-
-    @Test("a run that never reaches completion never reports the diagnostics moment")
-    @MainActor
-    func failingRunNeverReportsDiagnosticsMoment() async {
-        let probe = BlockingDiagnosticProbe()
-        let guidance = IsolatedGuidance()
-        let viewModel = NetworkDiagnosticsViewModel(
-            checks: [BlockingProbeDiagnosticCheck(id: .path, probe: probe)],
-            fingerprintMonitor: DisabledNetworkFingerprintMonitor(),
-            guidance: guidance.coordinator
-        )
-
-        #expect(viewModel.start())
-        await probe.waitForInvocationCount(1)
-        viewModel.cancel()
-        await viewModel.waitForCompletion()
-
-        #expect(viewModel.phase == .completed)
-        #expect(viewModel.conclusion == nil)
-        #expect(viewModel.endReason == .cancelled)
-        #expect(guidance.store.load().meaningfulCompletionCount == 0)
-        #expect(guidance.events.isEmpty)
-    }
-
-    @Test("system fingerprint monitor detects DNS settings changes without a path update")
-    func fingerprintDetectsSettingsOnlyChange() async throws {
-        let baseline = makeFingerprint(interfaceName: "en0", dnsHash: 1)
-        let path = NetworkPathFingerprint(
-            interfaceType: baseline.interfaceType,
-            interfaceName: baseline.interfaceName,
-            pathStatus: baseline.pathStatus
-        )
-        let settingsReader = SequencedFingerprintSettingsReader(dnsHashes: [1, 2], proxyHash: 7)
-        let monitor = SystemNetworkFingerprintMonitor(
-            settingsReader: settingsReader,
-            pathSource: FiniteNetworkPathFingerprintSource(values: [path]),
-            settingsPoller: ImmediateNetworkFingerprintSettingsPoller(),
-            settingsPollInterval: .milliseconds(10)
-        )
-
-        let optionalObservation = await monitor.observation()
-        let observation = try #require(optionalObservation)
-        var iterator = observation.changes.makeAsyncIterator()
-        let changedFingerprint = await iterator.next()
-
-        #expect(observation.baseline == baseline)
-        #expect(changedFingerprint?.interfaceName == "en0")
-        #expect(changedFingerprint?.dnsSettingsHash == 2)
-        #expect(changedFingerprint?.staticProxySettingsHash == 7)
-    }
-
-    @Test("system fingerprint monitor detects route changes without DNS or proxy changes")
-    func fingerprintDetectsRouteOnlyChange() async throws {
-        let baselinePath = NetworkPathFingerprint(
-            interfaceType: "ethernet",
-            interfaceName: "en7",
-            pathStatus: .satisfied
-        )
-        let routeSource = SequencedFingerprintRouteStateSource(values: [
-            NetworkFingerprintRouteState(
-                interfaceName: "en7",
-                interfaceIndex: 12,
-                gateway: "192.0.2.1",
-                addresses: ["192.0.2.10"],
-                subnets: ["255.255.255.0"]
-            ),
-            NetworkFingerprintRouteState(
-                interfaceName: "en7",
-                interfaceIndex: 12,
-                gateway: "192.0.2.254",
-                addresses: ["192.0.2.10"],
-                subnets: ["255.255.255.0"]
-            ),
-            NetworkFingerprintRouteState(
-                interfaceName: nil,
-                interfaceIndex: nil,
-                gateway: nil
-            ),
-        ])
-        let monitor = SystemNetworkFingerprintMonitor(
-            settingsReader: MutableFingerprintSettingsReader(dnsHash: 1, proxyHash: 7),
-            pathSource: FiniteNetworkPathFingerprintSource(values: [baselinePath]),
-            settingsPoller: FixedCountNetworkFingerprintSettingsPoller(count: 2),
-            routeStateSource: routeSource
-        )
-
-        let observation = try #require(await monitor.observation())
-        var iterator = observation.changes.makeAsyncIterator()
-        let change = await iterator.next()
-        let disappeared = await iterator.next()
-
-        #expect(observation.baseline.selectedGateway == "192.0.2.1")
-        #expect(change?.selectedGateway == "192.0.2.254")
-        #expect(disappeared?.selectedGateway == nil)
-        #expect(disappeared?.selectedInterfaceIndex == nil)
-        #expect(disappeared?.selectedInterfaceAddresses.isEmpty == true)
-    }
-
-    @Test("proxy fingerprint includes PAC and automatic discovery settings")
-    func proxyFingerprintIncludesAutomaticSettings() {
-        let baseline = SystemNetworkFingerprintSettingsReader.proxySettingsHash(for: [
-            "ProxyAutoConfigEnable": 0,
-            "ProxyAutoConfigURLString": "https://proxy.example/old.pac",
-            "ProxyAutoDiscoveryEnable": 0,
-        ])
-        let changedPAC = SystemNetworkFingerprintSettingsReader.proxySettingsHash(for: [
-            "ProxyAutoConfigEnable": 1,
-            "ProxyAutoConfigURLString": "https://proxy.example/new.pac",
-            "ProxyAutoDiscoveryEnable": 0,
-        ])
-        let changedDiscovery = SystemNetworkFingerprintSettingsReader.proxySettingsHash(for: [
-            "ProxyAutoConfigEnable": 0,
-            "ProxyAutoConfigURLString": "https://proxy.example/old.pac",
-            "ProxyAutoDiscoveryEnable": 1,
-        ])
-
-        #expect(changedPAC != baseline)
-        #expect(changedDiscovery != baseline)
-    }
-
-    @Test("VPN and TUN interface names are classified without VPN manager access")
-    func tunnelInterfaceClassification() {
-        let tunnels = NetworkTunnelInterfaceClassifier.tunnelInterfaces(from: [
-            "en0", "utun3", "ipsec0", "ppp0", "bridge0", "utun2"
-        ])
-
-        #expect(tunnels == ["ipsec0", "ppp0", "utun2", "utun3"])
-        #expect(NetworkTunnelInterfaceClassifier.routedTunnelInterface(
-            activeInterfaceName: "utun3",
-            tunnelInterfaces: tunnels
-        ) == "utun3")
-        #expect(NetworkTunnelInterfaceClassifier.routedTunnelInterface(
-            activeInterfaceName: "en0",
-            tunnelInterfaces: tunnels
-        ) == nil)
-    }
-
-    @Test("tunnel presence and routed interface changes are fingerprint changes")
-    func tunnelFingerprintChangesTriggerObservation() async throws {
-        let baselinePath = NetworkPathFingerprint(
-            interfaceType: "wifi",
-            interfaceName: "en0",
-            pathStatus: .satisfied,
-            tunnelInterfaces: [],
-            routedTunnelInterface: nil
-        )
-        let changedPath = NetworkPathFingerprint(
-            interfaceType: "wifi",
-            interfaceName: "en0",
-            pathStatus: .satisfied,
-            tunnelInterfaces: ["utun3"],
-            routedTunnelInterface: "utun3"
-        )
-        let monitor = SystemNetworkFingerprintMonitor(
-            settingsReader: MutableFingerprintSettingsReader(dnsHash: 1, proxyHash: 7),
-            pathSource: FiniteNetworkPathFingerprintSource(values: [baselinePath, changedPath]),
-            settingsPoller: SilentNetworkFingerprintSettingsPoller()
-        )
-
-        let observation = try #require(await monitor.observation())
-        var iterator = observation.changes.makeAsyncIterator()
-        let change = await iterator.next()
-
-        #expect(observation.baseline.tunnelInterfaces.isEmpty)
-        #expect(change?.tunnelInterfaces == ["utun3"])
-        #expect(change?.routedTunnelInterface == "utun3")
-    }
-
-    @Test("gateway target identity includes interface index and address state")
-    func fingerprintIncludesRouteAndAddressIdentity() {
-        let first = NetworkFingerprint(
-            interfaceType: "ethernet",
-            interfaceName: "en7",
-            pathStatus: .satisfied,
-            dnsSettingsHash: 1,
-            staticProxySettingsHash: 7,
-            selectedInterfaceIndex: 12,
-            selectedInterfaceAddresses: ["192.0.2.10"],
-            selectedInterfaceSubnets: ["255.255.255.0"],
-            selectedGateway: "192.0.2.1",
-            ipv4PrimaryServiceIdentity: "service-a",
-            ipv6PrimaryServiceIdentity: "service-v6"
-        )
-        let changedAddress = NetworkFingerprint(
-            interfaceType: "ethernet",
-            interfaceName: "en7",
-            pathStatus: .satisfied,
-            dnsSettingsHash: 1,
-            staticProxySettingsHash: 7,
-            selectedInterfaceIndex: 12,
-            selectedInterfaceAddresses: ["192.0.2.11"],
-            selectedInterfaceSubnets: ["255.255.255.0"],
-            selectedGateway: "192.0.2.1",
-            ipv4PrimaryServiceIdentity: "service-a",
-            ipv6PrimaryServiceIdentity: "service-v6"
-        )
-        let changedRoute = NetworkFingerprint(
-            interfaceType: "ethernet",
-            interfaceName: "en7",
-            pathStatus: .satisfied,
-            dnsSettingsHash: 1,
-            staticProxySettingsHash: 7,
-            selectedInterfaceIndex: 12,
-            selectedInterfaceAddresses: ["192.0.2.10"],
-            selectedInterfaceSubnets: ["255.255.255.0"],
-            selectedGateway: "192.0.2.254",
-            ipv4PrimaryServiceIdentity: "service-a",
-            ipv6PrimaryServiceIdentity: "service-v6"
-        )
-
-        #expect(first != changedAddress)
-        #expect(first != changedRoute)
-    }
-
-    @Test("fingerprint comparison ignores address and tunnel collection order")
-    func fingerprintNormalizesUnorderedCollections() {
-        let first = NetworkFingerprint(
-            interfaceType: "ethernet",
-            interfaceName: "en7",
-            pathStatus: .satisfied,
-            dnsSettingsHash: 1,
-            staticProxySettingsHash: 7,
-            tunnelInterfaces: ["utun3", "utun2"],
-            selectedInterfaceAddresses: ["192.0.2.11", "192.0.2.10"],
-            selectedInterfaceSubnets: ["255.255.255.0", "255.255.0.0"]
-        )
-        let reordered = NetworkFingerprint(
-            interfaceType: "ethernet",
-            interfaceName: "en7",
-            pathStatus: .satisfied,
-            dnsSettingsHash: 1,
-            staticProxySettingsHash: 7,
-            tunnelInterfaces: ["utun2", "utun3"],
-            selectedInterfaceAddresses: ["192.0.2.10", "192.0.2.11"],
-            selectedInterfaceSubnets: ["255.255.0.0", "255.255.255.0"]
-        )
-
-        #expect(first == reordered)
-        #expect(first.restartReason(comparedWith: reordered) == .path)
-    }
-
-    private func makeResults(
+    func makeResults(
         path: NetworkDiagnosticStatus,
         gateway: NetworkDiagnosticStatus = .normal,
         dns: NetworkDiagnosticStatus,
@@ -3495,7 +25,7 @@ struct NetworkDiagnosticsTests {
         ]
     }
 
-    private func makeFingerprint(
+    func makeFingerprint(
         interfaceName: String,
         dnsHash: UInt64
     ) -> NetworkFingerprint {
@@ -3508,7 +38,7 @@ struct NetworkDiagnosticsTests {
         )
     }
 
-    private func makeStubChecks(recorder: DiagnosticTestRecorder) -> [any DiagnosticCheck] {
+    func makeStubChecks(recorder: DiagnosticTestRecorder) -> [any DiagnosticCheck] {
         [.path, .dns, .proxy].map { id in
             StubDiagnosticCheck(
                 id: id,
@@ -3519,20 +49,20 @@ struct NetworkDiagnosticsTests {
     }
 }
 
-private struct AppLocalNetworkUsageDescription {
+struct AppLocalNetworkUsageDescription {
     let target: String
     let configuration: String
     let baseConfiguration: String
     let value: String
 }
 
-private func appLocalNetworkUsageDescriptions() throws -> [AppLocalNetworkUsageDescription] {
+func appLocalNetworkUsageDescriptions() throws -> [AppLocalNetworkUsageDescription] {
     let projectURL = try privacyCopyProjectURL()
     let project = try String(contentsOf: projectURL, encoding: .utf8)
     return try appLocalNetworkUsageDescriptions(from: project)
 }
 
-private func appLocalNetworkUsageDescriptions(
+func appLocalNetworkUsageDescriptions(
     from project: String
 ) throws -> [AppLocalNetworkUsageDescription] {
     let settingPrefix = "INFOPLIST_KEY_NSLocalNetworkUsageDescription = \""
@@ -3604,7 +134,7 @@ private func appLocalNetworkUsageDescriptions(
     }
 }
 
-private func nativeTarget(named name: String, in project: String) throws -> String {
+func nativeTarget(named name: String, in project: String) throws -> String {
     let marker = " /* \(name) */ = {\n\t\t\tisa = PBXNativeTarget;"
     let markerRange = try requiredRange(of: marker, in: project, context: "\(name) PBXNativeTarget")
     let lineStart = project[..<markerRange.lowerBound].lastIndex(of: "\n")
@@ -3615,7 +145,7 @@ private func nativeTarget(named name: String, in project: String) throws -> Stri
     return try pbxObject(id: identifier, in: project, context: "\(name) PBXNativeTarget")
 }
 
-private func pbxObject(id: String, in project: String, context: String) throws -> String {
+func pbxObject(id: String, in project: String, context: String) throws -> String {
     let header = "\t\t\(id) "
     let start: String.Index
     if project.hasPrefix(header) {
@@ -3633,7 +163,7 @@ private func pbxObject(id: String, in project: String, context: String) throws -
     return String(suffix[..<end])
 }
 
-private func buildConfigurationIDs(in configurationList: String) throws -> [String] {
+func buildConfigurationIDs(in configurationList: String) throws -> [String] {
     let start = try requiredRange(
         of: "buildConfigurations = (",
         in: configurationList,
@@ -3652,7 +182,7 @@ private func buildConfigurationIDs(in configurationList: String) throws -> [Stri
         }
 }
 
-private func pbxIdentifier(after prefix: String, in object: String, context: String) throws -> String {
+func pbxIdentifier(after prefix: String, in object: String, context: String) throws -> String {
     let value = try pbxValue(after: prefix, in: object, context: context)
     guard !value.isEmpty else {
         throw projectConfigurationError("\(context) must name a PBX object identifier.")
@@ -3660,28 +190,28 @@ private func pbxIdentifier(after prefix: String, in object: String, context: Str
     return value
 }
 
-private func pbxValue(after prefix: String, in object: String, context: String) throws -> String {
+func pbxValue(after prefix: String, in object: String, context: String) throws -> String {
     let valueStart = try requiredRange(of: prefix, in: object, context: context).upperBound
     let valueEnd = try requiredIndex(of: ";", in: object[valueStart...], context: "\(context) terminator")
     let value = object[valueStart..<valueEnd].trimmingCharacters(in: .whitespaces)
     return value.split(separator: " ").first.map(String.init) ?? ""
 }
 
-private func requiredMatch(_ candidates: [String], in value: String, context: String) throws -> String {
+func requiredMatch(_ candidates: [String], in value: String, context: String) throws -> String {
     guard let match = candidates.first(where: value.contains) else {
         throw projectConfigurationError("\(context) must reference OSS.xcconfig or PRO.xcconfig.")
     }
     return match
 }
 
-private func requiredRange(of needle: String, in value: String, context: String) throws -> Range<String.Index> {
+func requiredRange(of needle: String, in value: String, context: String) throws -> Range<String.Index> {
     guard let range = value.range(of: needle) else {
         throw projectConfigurationError("Missing \(context).")
     }
     return range
 }
 
-private func requiredIndex(
+func requiredIndex(
     of character: Character,
     in value: Substring,
     context: String
@@ -3692,7 +222,7 @@ private func requiredIndex(
     return index
 }
 
-private func projectConfigurationError(_ description: String) -> NSError {
+func projectConfigurationError(_ description: String) -> NSError {
     NSError(
         domain: "NetworkDiagnosticsTests",
         code: 1,
@@ -3700,7 +230,7 @@ private func projectConfigurationError(_ description: String) -> NSError {
     )
 }
 
-private func privacyCopyProjectURL() throws -> URL {
+func privacyCopyProjectURL() throws -> URL {
     var directory = URL(filePath: #filePath).deletingLastPathComponent()
     let marker = "WiFiLens/WiFiLens.xcodeproj/project.pbxproj"
 
@@ -3719,103 +249,7 @@ private func privacyCopyProjectURL() throws -> URL {
     )
 }
 
-private let privacyCopyProjectFixture = """
-\t\tOSS_TARGET /* WiFiLens */ = {
-\t\t\tisa = PBXNativeTarget;
-\t\t\tbuildConfigurationList = OSS_LIST /* Build configuration list for PBXNativeTarget \"WiFiLens\" */;
-\t\t\tname = WiFiLens;
-\t\t\tproductType = \"com.apple.product-type.application\";
-\t\t};
-\t\tPRO_TARGET /* WiFiLensPro */ = {
-\t\t\tisa = PBXNativeTarget;
-\t\t\tbuildConfigurationList = PRO_LIST /* Build configuration list for PBXNativeTarget \"WiFiLensPro\" */;
-\t\t\tname = WiFiLensPro;
-\t\t\tproductType = \"com.apple.product-type.application\";
-\t\t};
-\t\tDECOY_TARGET /* OtherApp */ = {
-\t\t\tisa = PBXNativeTarget;
-\t\t\tbuildConfigurationList = DECOY_LIST /* Build configuration list for PBXNativeTarget \"OtherApp\" */;
-\t\t\tname = OtherApp;
-\t\t\tproductType = \"com.apple.product-type.application\";
-\t\t};
-\t\tOSS_LIST /* Build configuration list */ = {
-\t\t\tisa = XCConfigurationList;
-\t\t\tbuildConfigurations = (
-\t\t\t\tOSS_DEBUG /* Debug */,
-\t\t\t\tOSS_RELEASE /* Release */,
-\t\t\t);
-\t\t};
-\t\tPRO_LIST /* Build configuration list */ = {
-\t\t\tisa = XCConfigurationList;
-\t\t\tbuildConfigurations = (
-\t\t\t\tPRO_DEBUG /* Debug */,
-\t\t\t\tPRO_RELEASE /* Release */,
-\t\t\t);
-\t\t};
-\t\tDECOY_LIST /* Build configuration list */ = {
-\t\t\tisa = XCConfigurationList;
-\t\t\tbuildConfigurations = (
-\t\t\t\tDECOY_DEBUG /* Debug */,
-\t\t\t\tDECOY_RELEASE /* Release */,
-\t\t\t);
-\t\t};
-\t\tOSS_DEBUG /* Debug */ = {
-\t\t\tisa = XCBuildConfiguration;
-\t\t\tbaseConfigurationReference = OSS_BASE /* OSS.xcconfig */;
-\t\t\tbuildSettings = {
-\t\t\t\tINFOPLIST_FILE = \"WiFiLens-Info.plist\";
-\t\t\t\tINFOPLIST_KEY_NSLocalNetworkUsageDescription = \"expected app copy\";
-\t\t\t};
-\t\t\tname = Debug;
-\t\t};
-\t\tOSS_RELEASE /* Release */ = {
-\t\t\tisa = XCBuildConfiguration;
-\t\t\tbaseConfigurationReference = OSS_BASE /* OSS.xcconfig */;
-\t\t\tbuildSettings = {
-\t\t\t\tINFOPLIST_FILE = \"WiFiLens-Info.plist\";
-\t\t\t\tINFOPLIST_KEY_NSLocalNetworkUsageDescription = \"expected app copy\";
-\t\t\t};
-\t\t\tname = Release;
-\t\t};
-\t\tPRO_DEBUG /* Debug */ = {
-\t\t\tisa = XCBuildConfiguration;
-\t\t\tbaseConfigurationReference = PRO_BASE /* PRO.xcconfig */;
-\t\t\tbuildSettings = {
-\t\t\t\tINFOPLIST_FILE = \"WiFiLensPro-Info.plist\";
-\t\t\t\tINFOPLIST_KEY_NSLocalNetworkUsageDescription = \"expected app copy\";
-\t\t\t};
-\t\t\tname = Debug;
-\t\t};
-\t\tPRO_RELEASE /* Release */ = {
-\t\t\tisa = XCBuildConfiguration;
-\t\t\tbaseConfigurationReference = PRO_BASE /* PRO.xcconfig */;
-\t\t\tbuildSettings = {
-\t\t\t\tINFOPLIST_FILE = \"WiFiLensPro-Info.plist\";
-\t\t\t\tINFOPLIST_KEY_NSLocalNetworkUsageDescription = \"expected app copy\";
-\t\t\t};
-\t\t\tname = Release;
-\t\t};
-\t\tDECOY_DEBUG /* Debug */ = {
-\t\t\tisa = XCBuildConfiguration;
-\t\t\tbaseConfigurationReference = DECOY_BASE /* OSS.xcconfig */;
-\t\t\tbuildSettings = {
-\t\t\t\tINFOPLIST_FILE = \"Other-Info.plist\";
-\t\t\t\tINFOPLIST_KEY_NSLocalNetworkUsageDescription = \"decoy copy\";
-\t\t\t};
-\t\t\tname = Debug;
-\t\t};
-\t\tDECOY_RELEASE /* Release */ = {
-\t\t\tisa = XCBuildConfiguration;
-\t\t\tbaseConfigurationReference = DECOY_BASE /* OSS.xcconfig */;
-\t\t\tbuildSettings = {
-\t\t\t\tINFOPLIST_FILE = \"Other-Info.plist\";
-\t\t\t\tINFOPLIST_KEY_NSLocalNetworkUsageDescription = \"decoy copy\";
-\t\t\t};
-\t\t\tname = Release;
-\t\t};
-"""
-
-private struct StubPathSource: NetworkPathChecking {
+struct StubPathSource: NetworkPathChecking {
     let state: NetworkPathState?
 
     init(_ state: NetworkPathState?) {
@@ -3827,9 +261,9 @@ private struct StubPathSource: NetworkPathChecking {
     }
 }
 
-private actor ConcurrentDNSResolver: DNSResolving {
+actor ConcurrentDNSResolver: DNSResolving {
     private(set) var maximumInFlight = 0
-    private var inFlight = 0
+    var inFlight = 0
 
     func resolve(host: String, timeout: Duration) async -> DNSResolutionOutcome {
         inFlight += 1
@@ -3840,9 +274,9 @@ private actor ConcurrentDNSResolver: DNSResolving {
     }
 }
 
-private actor ConcurrentControlLoader: ControlEndpointLoading {
+actor ConcurrentControlLoader: ControlEndpointLoading {
     private(set) var maximumInFlight = 0
-    private var inFlight = 0
+    var inFlight = 0
 
     func load(url: URL, timeout: Duration) async -> ControlEndpointLoadResult {
         inFlight += 1
@@ -3856,7 +290,7 @@ private actor ConcurrentControlLoader: ControlEndpointLoading {
     }
 }
 
-private struct MetricsControlLoader: ControlEndpointLoading {
+struct MetricsControlLoader: ControlEndpointLoading {
     let metrics: ControlEndpointMetrics
 
     func load(url: URL, timeout: Duration) async -> ControlEndpointLoadResult {
@@ -3867,9 +301,9 @@ private struct MetricsControlLoader: ControlEndpointLoading {
     }
 }
 
-private actor BudgetAwareDiagnosticProbe {
-    private var didStart = false
-    private var invocationContinuation: CheckedContinuation<Void, Never>?
+actor BudgetAwareDiagnosticProbe {
+    var didStart = false
+    var invocationContinuation: CheckedContinuation<Void, Never>?
     private(set) var wasCancelled = false
 
     func run() async {
@@ -3897,10 +331,10 @@ private actor BudgetAwareDiagnosticProbe {
     }
 }
 
-private actor ManualDiagnosticClock: DiagnosticClock {
+actor ManualDiagnosticClock: DiagnosticClock {
     let origin: ContinuousClock.Instant
-    private var current: ContinuousClock.Instant
-    private var sleepers: [UUID: (deadline: ContinuousClock.Instant, continuation: CheckedContinuation<Void, Never>)] = [:]
+    var current: ContinuousClock.Instant
+    var sleepers: [UUID: (deadline: ContinuousClock.Instant, continuation: CheckedContinuation<Void, Never>)] = [:]
 
     init(origin: ContinuousClock.Instant = ContinuousClock.now) {
         self.origin = origin
@@ -3946,13 +380,13 @@ private actor ManualDiagnosticClock: DiagnosticClock {
         }
     }
 
-    private func cancelSleep(id: UUID) {
+    func cancelSleep(id: UUID) {
         guard let sleeper = sleepers.removeValue(forKey: id) else { return }
         sleeper.continuation.resume()
     }
 }
 
-private actor OptionalBoolRecorder {
+actor OptionalBoolRecorder {
     private(set) var value: Bool?
 
     func record(_ value: Bool) {
@@ -3960,7 +394,7 @@ private actor OptionalBoolRecorder {
     }
 }
 
-private struct BudgetAwareDiagnosticCheck: DiagnosticCheck {
+struct BudgetAwareDiagnosticCheck: DiagnosticCheck {
     let id: NetworkDiagnosticCheckID = .path
     let probe: BudgetAwareDiagnosticProbe
 
@@ -3970,9 +404,9 @@ private struct BudgetAwareDiagnosticCheck: DiagnosticCheck {
     }
 }
 
-private actor StubDNSResolver: DNSResolving {
-    private var outcomes: [DNSResolutionOutcome]
-    private var outcomeIndex = 0
+actor StubDNSResolver: DNSResolving {
+    var outcomes: [DNSResolutionOutcome]
+    var outcomeIndex = 0
     private(set) var invocationCount = 0
 
     init(_ outcome: DNSResolutionOutcome) {
@@ -3991,7 +425,7 @@ private actor StubDNSResolver: DNSResolving {
     }
 }
 
-private struct MappingDNSResolver: DNSResolving {
+struct MappingDNSResolver: DNSResolving {
     let outcomes: [String: DNSResolutionOutcome]
 
     func resolve(host: String, timeout: Duration) async -> DNSResolutionOutcome {
@@ -3999,8 +433,8 @@ private struct MappingDNSResolver: DNSResolving {
     }
 }
 
-private actor CancellationAwareDNSResolver: DNSResolving {
-    private var invocationWaiter: CheckedContinuation<Void, Never>?
+actor CancellationAwareDNSResolver: DNSResolving {
+    var invocationWaiter: CheckedContinuation<Void, Never>?
     private(set) var invocationCount = 0
 
     func resolve(host: String, timeout: Duration) async -> DNSResolutionOutcome {
@@ -4017,7 +451,7 @@ private actor CancellationAwareDNSResolver: DNSResolving {
     }
 }
 
-private struct StubControlLoader: ControlEndpointLoading {
+struct StubControlLoader: ControlEndpointLoading {
     let httpsStatus: Int?
     let httpsErrorCode: String?
     let httpStatus: Int?
@@ -4050,7 +484,7 @@ private struct StubControlLoader: ControlEndpointLoading {
     }
 }
 
-private struct StubIPv6Loader: IPv6ControlEndpointLoading {
+struct StubIPv6Loader: IPv6ControlEndpointLoading {
     let outcome: IPv6ControlEndpointLoadOutcome
 
     init(_ outcome: IPv6ControlEndpointLoadOutcome) {
@@ -4062,7 +496,7 @@ private struct StubIPv6Loader: IPv6ControlEndpointLoading {
     }
 }
 
-private struct StubGlobalIPv6AddressSource: GlobalIPv6AddressSourcing {
+struct StubGlobalIPv6AddressSource: GlobalIPv6AddressSourcing {
     let hasAddress: Bool
 
     func hasGlobalIPv6Address() -> Bool {
@@ -4070,7 +504,7 @@ private struct StubGlobalIPv6AddressSource: GlobalIPv6AddressSourcing {
     }
 }
 
-private struct StubIPv6AddressResolver: IPv6AddressResolving {
+struct StubIPv6AddressResolver: IPv6AddressResolving {
     let addresses: [String]
 
     func resolveAAAA(host: String, timeout: Duration) async -> [String] {
@@ -4078,7 +512,7 @@ private struct StubIPv6AddressResolver: IPv6AddressResolving {
     }
 }
 
-private actor RecordingIPv6AddressResolver: IPv6AddressResolving {
+actor RecordingIPv6AddressResolver: IPv6AddressResolving {
     let addresses: [String]
     private(set) var hosts: [String] = []
 
@@ -4092,14 +526,14 @@ private actor RecordingIPv6AddressResolver: IPv6AddressResolving {
     }
 }
 
-private struct IPv6LoaderTestRequest: Equatable, Sendable {
+struct IPv6LoaderTestRequest: Equatable, Sendable {
     let url: URL
     let ipv6Address: String
     let serverName: String
     let timeout: Duration
 }
 
-private actor IPv6LoaderTestRecorder {
+actor IPv6LoaderTestRecorder {
     private(set) var requests: [IPv6LoaderTestRequest] = []
 
     func record(_ request: IPv6LoaderTestRequest) {
@@ -4107,7 +541,7 @@ private actor IPv6LoaderTestRecorder {
     }
 }
 
-private struct RecordingIPv6HTTPSConnector: IPv6HTTPSConnecting {
+struct RecordingIPv6HTTPSConnector: IPv6HTTPSConnecting {
     let succeeds: Bool
     let recorder: IPv6LoaderTestRecorder
 
@@ -4127,7 +561,7 @@ private struct RecordingIPv6HTTPSConnector: IPv6HTTPSConnecting {
     }
 }
 
-private struct StubIPv6HTTPSConnector: IPv6HTTPSConnecting {
+struct StubIPv6HTTPSConnector: IPv6HTTPSConnecting {
     let succeeds: Bool
 
     func load(
@@ -4140,7 +574,7 @@ private struct StubIPv6HTTPSConnector: IPv6HTTPSConnecting {
     }
 }
 
-private actor SequencedIPv6HTTPSConnector: IPv6HTTPSConnecting {
+actor SequencedIPv6HTTPSConnector: IPv6HTTPSConnecting {
     let successfulAddress: String
     private(set) var addresses: [String] = []
     private(set) var timeouts: [Duration] = []
@@ -4161,7 +595,7 @@ private actor SequencedIPv6HTTPSConnector: IPv6HTTPSConnecting {
     }
 }
 
-private actor ControlEndpointTestRecorder {
+actor ControlEndpointTestRecorder {
     private(set) var urls: [String] = []
     private(set) var timeouts: [Duration] = []
 
@@ -4171,7 +605,7 @@ private actor ControlEndpointTestRecorder {
     }
 }
 
-private struct StubProxyConfigurationResolver: ProxyConfigurationResolving {
+struct StubProxyConfigurationResolver: ProxyConfigurationResolving {
     let storedResolutions: [ProxyResolutionDirective]
 
     init(_ resolution: ProxyResolutionDirective) {
@@ -4187,13 +621,13 @@ private struct StubProxyConfigurationResolver: ProxyConfigurationResolving {
     }
 }
 
-private struct PACResolutionTestRequest: Equatable, Sendable {
+struct PACResolutionTestRequest: Equatable, Sendable {
     let pacURL: URL
     let targetURL: URL
     let timeout: Duration
 }
 
-private actor PACResolutionTestRecorder {
+actor PACResolutionTestRecorder {
     private(set) var requests: [PACResolutionTestRequest] = []
 
     func record(pacURL: URL, targetURL: URL, timeout: Duration) {
@@ -4201,14 +635,14 @@ private actor PACResolutionTestRecorder {
     }
 }
 
-private struct PACCallbackTestRequest: Equatable, Sendable {
+struct PACCallbackTestRequest: Equatable, Sendable {
     let source: PACSource
     let targetURL: URL
 }
 
-private final class ControlledPACCallbackExecution: PACCallbackExecution, @unchecked Sendable {
-    private let lock = NSLock()
-    private var cancelled = false
+final class ControlledPACCallbackExecution: PACCallbackExecution, @unchecked Sendable {
+    let lock = NSLock()
+    var cancelled = false
 
     var isCancelled: Bool {
         lock.withLock { cancelled }
@@ -4219,12 +653,12 @@ private final class ControlledPACCallbackExecution: PACCallbackExecution, @unche
     }
 }
 
-private final class ControlledPACCallbackExecutor: PACCallbackExecuting, @unchecked Sendable {
-    private let lock = NSLock()
-    private let requestStream: AsyncStream<PACCallbackTestRequest>
-    private let requestContinuation: AsyncStream<PACCallbackTestRequest>.Continuation
-    private var callback: (@Sendable (PACCallbackOutcome) -> Void)?
-    private var execution: ControlledPACCallbackExecution?
+final class ControlledPACCallbackExecutor: PACCallbackExecuting, @unchecked Sendable {
+    let lock = NSLock()
+    let requestStream: AsyncStream<PACCallbackTestRequest>
+    let requestContinuation: AsyncStream<PACCallbackTestRequest>.Continuation
+    var callback: (@Sendable (PACCallbackOutcome) -> Void)?
+    var execution: ControlledPACCallbackExecution?
 
     init() {
         let pair = AsyncStream<PACCallbackTestRequest>.makeStream()
@@ -4267,7 +701,7 @@ private final class ControlledPACCallbackExecutor: PACCallbackExecuting, @unchec
     }
 }
 
-private struct StubPACResolver: PACResolving {
+struct StubPACResolver: PACResolving {
     let resolution: ProxyCandidateResolution
     let recorder: PACResolutionTestRecorder?
 
@@ -4296,9 +730,9 @@ private struct StubPACResolver: PACResolving {
     }
 }
 
-private actor CancellationIgnoringPACResolver: PACResolving {
-    private var invocationWaiters: [CheckedContinuation<Void, Never>] = []
-    private var hasInvoked = false
+actor CancellationIgnoringPACResolver: PACResolving {
+    var invocationWaiters: [CheckedContinuation<Void, Never>] = []
+    var hasInvoked = false
 
     func resolve(
         pacURL: URL,
@@ -4318,7 +752,7 @@ private actor CancellationIgnoringPACResolver: PACResolving {
     }
 }
 
-private struct StubProxyResolver: ProxyResolving {
+struct StubProxyResolver: ProxyResolving {
     let resolution: ProxyCandidateResolution
 
     init(_ resolution: EffectiveProxy) {
@@ -4339,7 +773,7 @@ private struct StubProxyResolver: ProxyResolving {
     }
 }
 
-private struct StubProxyTunnelStateReader: ProxyTunnelStateReading {
+struct StubProxyTunnelStateReader: ProxyTunnelStateReading {
     let state: ProxyTunnelState?
 
     init(
@@ -4356,10 +790,10 @@ private struct StubProxyTunnelStateReader: ProxyTunnelStateReading {
     }
 }
 
-private final class TunnelTimeoutCancellationProbe: @unchecked Sendable {
-    private let lock = NSLock()
-    private var parkedContinuation: CheckedContinuation<Void, Never>?
-    private var cancelled = false
+final class TunnelTimeoutCancellationProbe: @unchecked Sendable {
+    let lock = NSLock()
+    var parkedContinuation: CheckedContinuation<Void, Never>?
+    var cancelled = false
 
     var didCancel: Bool {
         lock.lock()
@@ -4389,9 +823,9 @@ private final class TunnelTimeoutCancellationProbe: @unchecked Sendable {
     }
 }
 
-private final class LockedFlag: @unchecked Sendable {
-    private let lock = NSLock()
-    private var value = false
+final class LockedFlag: @unchecked Sendable {
+    let lock = NSLock()
+    var value = false
 
     var isSet: Bool {
         lock.lock()
@@ -4406,7 +840,7 @@ private final class LockedFlag: @unchecked Sendable {
     }
 }
 
-private actor ProxyResolutionTestRecorder {
+actor ProxyResolutionTestRecorder {
     private(set) var urls: [String] = []
 
     func record(_ url: URL) {
@@ -4414,7 +848,7 @@ private actor ProxyResolutionTestRecorder {
     }
 }
 
-private struct RecordingProxyResolver: ProxyResolving {
+struct RecordingProxyResolver: ProxyResolving {
     let resolutions: [URL: ProxyCandidateResolution]
     let defaultResolution: ProxyCandidateResolution?
     let recorder: ProxyResolutionTestRecorder?
@@ -4445,9 +879,9 @@ private struct RecordingProxyResolver: ProxyResolving {
     }
 }
 
-private actor CancellationIgnoringProxyResolver: ProxyResolving {
+actor CancellationIgnoringProxyResolver: ProxyResolving {
     private(set) var urls: [String] = []
-    private var invocationWaiters: [CheckedContinuation<Void, Never>] = []
+    var invocationWaiters: [CheckedContinuation<Void, Never>] = []
 
     func resolve(for url: URL) async -> ProxyCandidateResolution {
         urls.append(url.absoluteString)
@@ -4465,7 +899,7 @@ private actor CancellationIgnoringProxyResolver: ProxyResolving {
     }
 }
 
-private func proxyDictionary(
+func proxyDictionary(
     type: CFString,
     host: String? = nil,
     port: Int? = nil
@@ -4477,12 +911,12 @@ private func proxyDictionary(
     return dictionary
 }
 
-private struct ProxyEgressTestRequest: Equatable, Sendable {
+struct ProxyEgressTestRequest: Equatable, Sendable {
     let url: URL
     let proxy: EffectiveProxy
 }
 
-private actor ProxyEgressTestRecorder {
+actor ProxyEgressTestRecorder {
     private(set) var requests: [ProxyEgressTestRequest] = []
 
     func record(url: URL, proxy: EffectiveProxy) {
@@ -4490,7 +924,7 @@ private actor ProxyEgressTestRecorder {
     }
 }
 
-private struct StubProxyEgressLoader: ProxyEgressLoading {
+struct StubProxyEgressLoader: ProxyEgressLoading {
     let statusCode: Int?
     let errorCode: String?
     let recorder: ProxyEgressTestRecorder?
@@ -4515,11 +949,11 @@ private struct StubProxyEgressLoader: ProxyEgressLoading {
     }
 }
 
-private actor SequencedProxyEgressLoader: ProxyEgressLoading {
+actor SequencedProxyEgressLoader: ProxyEgressLoading {
     let responses: [ProxyEgressResponse]
     let delays: [Duration]
     private(set) var timeouts: [Duration] = []
-    private var invocationCount = 0
+    var invocationCount = 0
 
     init(responses: [ProxyEgressResponse], delays: [Duration]) {
         self.responses = responses
@@ -4544,10 +978,10 @@ private actor SequencedProxyEgressLoader: ProxyEgressLoading {
     }
 }
 
-private final class SequencedProxyCheckClock: ProxyCheckClock, @unchecked Sendable {
-    private let lock = NSLock()
-    private let instants: [ContinuousClock.Instant]
-    private var index = 0
+final class SequencedProxyCheckClock: ProxyCheckClock, @unchecked Sendable {
+    let lock = NSLock()
+    let instants: [ContinuousClock.Instant]
+    var index = 0
 
     init(offsets: [Duration]) {
         let origin = ContinuousClock().now
@@ -4563,8 +997,8 @@ private final class SequencedProxyCheckClock: ProxyCheckClock, @unchecked Sendab
     }
 }
 
-private actor CancellationIgnoringProxyEgressLoader: ProxyEgressLoading {
-    private var invocationWaiters: [CheckedContinuation<Void, Never>] = []
+actor CancellationIgnoringProxyEgressLoader: ProxyEgressLoading {
+    var invocationWaiters: [CheckedContinuation<Void, Never>] = []
     private(set) var invocationCount = 0
 
     func load(
@@ -4587,7 +1021,7 @@ private actor CancellationIgnoringProxyEgressLoader: ProxyEgressLoading {
     }
 }
 
-private struct StubProxyConnector: ProxyEndpointConnecting {
+struct StubProxyConnector: ProxyEndpointConnecting {
     let reachable: Bool
 
     func canConnect(to endpoint: ProxyEndpoint, timeout: Duration) async -> Bool {
@@ -4595,7 +1029,7 @@ private struct StubProxyConnector: ProxyEndpointConnecting {
     }
 }
 
-private struct EndpointSelectiveProxyConnector: ProxyEndpointConnecting {
+struct EndpointSelectiveProxyConnector: ProxyEndpointConnecting {
     let reachableHosts: Set<String>
 
     func canConnect(to endpoint: ProxyEndpoint, timeout: Duration) async -> Bool {
@@ -4603,7 +1037,7 @@ private struct EndpointSelectiveProxyConnector: ProxyEndpointConnecting {
     }
 }
 
-private actor ProxyConnectorTestRecorder {
+actor ProxyConnectorTestRecorder {
     private(set) var endpoints: [ProxyEndpoint] = []
 
     func record(_ endpoint: ProxyEndpoint) {
@@ -4611,11 +1045,11 @@ private actor ProxyConnectorTestRecorder {
     }
 }
 
-private actor SequencedProxyConnector: ProxyEndpointConnecting {
+actor SequencedProxyConnector: ProxyEndpointConnecting {
     let outcomes: [Bool]
     private(set) var endpoints: [ProxyEndpoint] = []
     private(set) var timeouts: [Duration] = []
-    private var invocationCount = 0
+    var invocationCount = 0
 
     init(outcomes: [Bool]) {
         self.outcomes = outcomes
@@ -4630,8 +1064,8 @@ private actor SequencedProxyConnector: ProxyEndpointConnecting {
     }
 }
 
-private actor CancellationIgnoringProxyConnector: ProxyEndpointConnecting {
-    private var invocationWaiters: [CheckedContinuation<Void, Never>] = []
+actor CancellationIgnoringProxyConnector: ProxyEndpointConnecting {
+    var invocationWaiters: [CheckedContinuation<Void, Never>] = []
     private(set) var endpoints: [ProxyEndpoint] = []
 
     func canConnect(to endpoint: ProxyEndpoint, timeout: Duration) async -> Bool {
@@ -4650,7 +1084,7 @@ private actor CancellationIgnoringProxyConnector: ProxyEndpointConnecting {
     }
 }
 
-private struct RecordingProxyConnector: ProxyEndpointConnecting {
+struct RecordingProxyConnector: ProxyEndpointConnecting {
     let reachable: Bool
     let recorder: ProxyConnectorTestRecorder
 
@@ -4660,7 +1094,7 @@ private struct RecordingProxyConnector: ProxyEndpointConnecting {
     }
 }
 
-private actor DiagnosticTestRecorder {
+actor DiagnosticTestRecorder {
     private(set) var values: [NetworkDiagnosticCheckID] = []
 
     func record(_ value: NetworkDiagnosticCheckID) {
@@ -4672,17 +1106,17 @@ private actor DiagnosticTestRecorder {
 /// sink. Keeps diagnostics view-model tests hermetic (no real UserDefaults,
 /// no Launch Services queries).
 @MainActor
-private final class IsolatedGuidance {
+final class IsolatedGuidance {
     let store: InMemoryGuidanceStateStore
     let coordinator: GuidanceCoordinator
-    private let eventBox: EventBox
+    let eventBox: DiagnosticGuidanceEventBox
 
     var events: [GuidanceEvent] { eventBox.events }
 
     init(completionCount: Int = 0) {
         let calendar = Self.makeCalendar()
         let now = calendar.date(from: DateComponents(year: 2026, month: 8, day: 5, hour: 12))!
-        let box = EventBox()
+        let box = DiagnosticGuidanceEventBox()
         store = InMemoryGuidanceStateStore(initial: GuidanceState(
             activeDays: ["2026-07-01", "2026-07-02"],
             meaningfulCompletionCount: completionCount
@@ -4711,11 +1145,11 @@ private final class IsolatedGuidance {
 }
 
 @MainActor
-private final class EventBox {
+final class DiagnosticGuidanceEventBox {
     var events: [GuidanceEvent] = []
 }
 
-private struct StubDiagnosticCheck: DiagnosticCheck {
+struct StubDiagnosticCheck: DiagnosticCheck {
     let id: NetworkDiagnosticCheckID
     let result: NetworkDiagnosticResult
     let recorder: DiagnosticTestRecorder
@@ -4726,10 +1160,10 @@ private struct StubDiagnosticCheck: DiagnosticCheck {
     }
 }
 
-private actor RestartableDiagnosticProbe {
-    private let blockFirstInvocation: Bool
-    private var firstInvocationContinuation: CheckedContinuation<Void, Never>?
-    private var invocationWaiters: [(count: Int, continuation: CheckedContinuation<Void, Never>)] = []
+actor RestartableDiagnosticProbe {
+    let blockFirstInvocation: Bool
+    var firstInvocationContinuation: CheckedContinuation<Void, Never>?
+    var invocationWaiters: [(count: Int, continuation: CheckedContinuation<Void, Never>)] = []
     private(set) var invocationCount = 0
 
     init(blockFirstInvocation: Bool = false) {
@@ -4768,7 +1202,7 @@ private actor RestartableDiagnosticProbe {
     }
 }
 
-private struct ProbeDiagnosticCheck: DiagnosticCheck {
+struct ProbeDiagnosticCheck: DiagnosticCheck {
     let id: NetworkDiagnosticCheckID
     let result: NetworkDiagnosticResult
     let rerunPolicy: DiagnosticCheckRerunPolicy
@@ -4780,9 +1214,9 @@ private struct ProbeDiagnosticCheck: DiagnosticCheck {
     }
 }
 
-private actor BlockingDiagnosticProbe {
-    private var continuations: [Int: CheckedContinuation<Void, Never>] = [:]
-    private var invocationWaiters: [(count: Int, continuation: CheckedContinuation<Void, Never>)] = []
+actor BlockingDiagnosticProbe {
+    var continuations: [Int: CheckedContinuation<Void, Never>] = [:]
+    var invocationWaiters: [(count: Int, continuation: CheckedContinuation<Void, Never>)] = []
     private(set) var invocationCount = 0
 
     func run() async {
@@ -4811,7 +1245,7 @@ private actor BlockingDiagnosticProbe {
     }
 }
 
-private struct BlockingProbeDiagnosticCheck: DiagnosticCheck {
+struct BlockingProbeDiagnosticCheck: DiagnosticCheck {
     let id: NetworkDiagnosticCheckID
     let probe: BlockingDiagnosticProbe
 
@@ -4821,9 +1255,9 @@ private struct BlockingProbeDiagnosticCheck: DiagnosticCheck {
     }
 }
 
-private actor CancellationIgnoringDiagnosticProbe {
-    private var continuations: [Int: CheckedContinuation<Void, Never>] = [:]
-    private var invocationWaiters: [(count: Int, continuation: CheckedContinuation<Void, Never>)] = []
+actor CancellationIgnoringDiagnosticProbe {
+    var continuations: [Int: CheckedContinuation<Void, Never>] = [:]
+    var invocationWaiters: [(count: Int, continuation: CheckedContinuation<Void, Never>)] = []
     private(set) var invocationCount = 0
 
     func run() async {
@@ -4850,7 +1284,7 @@ private actor CancellationIgnoringDiagnosticProbe {
     }
 }
 
-private struct CancellationIgnoringProbeDiagnosticCheck: DiagnosticCheck {
+struct CancellationIgnoringProbeDiagnosticCheck: DiagnosticCheck {
     let id: NetworkDiagnosticCheckID = .path
     let probe: CancellationIgnoringDiagnosticProbe
 
@@ -4860,11 +1294,11 @@ private struct CancellationIgnoringProbeDiagnosticCheck: DiagnosticCheck {
     }
 }
 
-private actor ControlledNetworkFingerprintMonitor: NetworkFingerprintMonitoring {
+actor ControlledNetworkFingerprintMonitor: NetworkFingerprintMonitoring {
     let initial: NetworkFingerprint
-    private var lastFingerprint: NetworkFingerprint
-    private var continuation: AsyncStream<NetworkFingerprint>.Continuation?
-    private var pending: [NetworkFingerprint] = []
+    var lastFingerprint: NetworkFingerprint
+    var continuation: AsyncStream<NetworkFingerprint>.Continuation?
+    var pending: [NetworkFingerprint] = []
 
     init(initial: NetworkFingerprint) {
         self.initial = initial
@@ -4887,7 +1321,7 @@ private actor ControlledNetworkFingerprintMonitor: NetworkFingerprintMonitoring 
         continuation.yield(fingerprint)
     }
 
-    private func install(_ continuation: AsyncStream<NetworkFingerprint>.Continuation) {
+    func install(_ continuation: AsyncStream<NetworkFingerprint>.Continuation) {
         self.continuation = continuation
         for fingerprint in pending {
             continuation.yield(fingerprint)
@@ -4896,8 +1330,8 @@ private actor ControlledNetworkFingerprintMonitor: NetworkFingerprintMonitoring 
     }
 }
 
-private actor CancellationIgnoringNetworkFingerprintMonitor: NetworkFingerprintMonitoring {
-    private var continuation: CheckedContinuation<NetworkFingerprintObservation?, Never>?
+actor CancellationIgnoringNetworkFingerprintMonitor: NetworkFingerprintMonitoring {
+    var continuation: CheckedContinuation<NetworkFingerprintObservation?, Never>?
 
     func observation() async -> NetworkFingerprintObservation? {
         await withCheckedContinuation { continuation in
@@ -4917,13 +1351,13 @@ private actor CancellationIgnoringNetworkFingerprintMonitor: NetworkFingerprintM
     }
 }
 
-private struct SilentNetworkPathFingerprintSource: NetworkPathFingerprintSourcing {
+struct SilentNetworkPathFingerprintSource: NetworkPathFingerprintSourcing {
     func pathFingerprintChanges() -> AsyncStream<NetworkPathFingerprint> {
         AsyncStream { $0.finish() }
     }
 }
 
-private struct FiniteNetworkPathFingerprintSource: NetworkPathFingerprintSourcing {
+struct FiniteNetworkPathFingerprintSource: NetworkPathFingerprintSourcing {
     let values: [NetworkPathFingerprint]
 
     func pathFingerprintChanges() -> AsyncStream<NetworkPathFingerprint> {
@@ -4934,10 +1368,10 @@ private struct FiniteNetworkPathFingerprintSource: NetworkPathFingerprintSourcin
     }
 }
 
-private final class BufferedGapNetworkPathFingerprintSource: NetworkPathFingerprintSourcing, @unchecked Sendable {
-    private let lock = NSLock()
-    private let values: [NetworkPathFingerprint]
-    private var requestCount = 0
+final class BufferedGapNetworkPathFingerprintSource: NetworkPathFingerprintSourcing, @unchecked Sendable {
+    let lock = NSLock()
+    let values: [NetworkPathFingerprint]
+    var requestCount = 0
 
     init(values: [NetworkPathFingerprint]) {
         self.values = values
@@ -4956,7 +1390,7 @@ private final class BufferedGapNetworkPathFingerprintSource: NetworkPathFingerpr
     }
 }
 
-private struct ImmediateNetworkFingerprintSettingsPoller: NetworkFingerprintSettingsPolling {
+struct ImmediateNetworkFingerprintSettingsPoller: NetworkFingerprintSettingsPolling {
     func ticks(every interval: Duration) -> AsyncStream<Void> {
         AsyncStream { continuation in
             continuation.yield(())
@@ -4965,7 +1399,7 @@ private struct ImmediateNetworkFingerprintSettingsPoller: NetworkFingerprintSett
     }
 }
 
-private struct FixedCountNetworkFingerprintSettingsPoller: NetworkFingerprintSettingsPolling {
+struct FixedCountNetworkFingerprintSettingsPoller: NetworkFingerprintSettingsPolling {
     let count: Int
 
     func ticks(every interval: Duration) -> AsyncStream<Void> {
@@ -4978,16 +1412,16 @@ private struct FixedCountNetworkFingerprintSettingsPoller: NetworkFingerprintSet
     }
 }
 
-private struct SilentNetworkFingerprintSettingsPoller: NetworkFingerprintSettingsPolling {
+struct SilentNetworkFingerprintSettingsPoller: NetworkFingerprintSettingsPolling {
     func ticks(every interval: Duration) -> AsyncStream<Void> {
         AsyncStream { $0.finish() }
     }
 }
 
-private final class MutableFingerprintSettingsReader: NetworkFingerprintSettingsReading, @unchecked Sendable {
-    private let lock = NSLock()
-    private var dnsHash: UInt64
-    private let proxyHash: UInt64
+final class MutableFingerprintSettingsReader: NetworkFingerprintSettingsReading, @unchecked Sendable {
+    let lock = NSLock()
+    var dnsHash: UInt64
+    let proxyHash: UInt64
 
     init(dnsHash: UInt64, proxyHash: UInt64) {
         self.dnsHash = dnsHash
@@ -5007,9 +1441,9 @@ private final class MutableFingerprintSettingsReader: NetworkFingerprintSettings
     }
 }
 
-private actor SequencedFingerprintRouteStateSource: NetworkFingerprintRouteStateSourcing {
-    private let values: [NetworkFingerprintRouteState]
-    private var index = 0
+actor SequencedFingerprintRouteStateSource: NetworkFingerprintRouteStateSourcing {
+    let values: [NetworkFingerprintRouteState]
+    var index = 0
 
     init(values: [NetworkFingerprintRouteState]) {
         precondition(!values.isEmpty)
@@ -5022,11 +1456,11 @@ private actor SequencedFingerprintRouteStateSource: NetworkFingerprintRouteState
     }
 }
 
-private final class SequencedFingerprintSettingsReader: NetworkFingerprintSettingsReading, @unchecked Sendable {
-    private let lock = NSLock()
-    private let dnsHashes: [UInt64]
-    private let proxyHash: UInt64
-    private var dnsIndex = 0
+final class SequencedFingerprintSettingsReader: NetworkFingerprintSettingsReading, @unchecked Sendable {
+    let lock = NSLock()
+    let dnsHashes: [UInt64]
+    let proxyHash: UInt64
+    var dnsIndex = 0
 
     init(dnsHashes: [UInt64], proxyHash: UInt64) {
         self.dnsHashes = dnsHashes
@@ -5047,7 +1481,7 @@ private final class SequencedFingerprintSettingsReader: NetworkFingerprintSettin
     }
 }
 
-private struct StubNetworkInterfaceSource: NetworkInterfaceInfoSourcing {
+struct StubNetworkInterfaceSource: NetworkInterfaceInfoSourcing {
     let interface: NetworkInterfaceInfo?
 
     func currentInterface() async -> NetworkInterfaceInfo? {
@@ -5055,7 +1489,7 @@ private struct StubNetworkInterfaceSource: NetworkInterfaceInfoSourcing {
     }
 }
 
-private actor RecordingGatewayPingProcessRunner: GatewayPingProcessRunning {
+actor RecordingGatewayPingProcessRunner: GatewayPingProcessRunning {
     let latency: Double?
     private(set) var executablePath: String?
     private(set) var arguments: [String] = []
@@ -5073,7 +1507,7 @@ private actor RecordingGatewayPingProcessRunner: GatewayPingProcessRunning {
     func cancel() async {}
 }
 
-private actor RecordingDiagnosticGatewayMeasurer: DiagnosticGatewayMeasuring {
+actor RecordingDiagnosticGatewayMeasurer: DiagnosticGatewayMeasuring {
     private(set) var targets: [DiagnosticGatewayTarget] = []
 
     func measure(target: DiagnosticGatewayTarget) async -> GatewayLatencyResult {
@@ -5086,9 +1520,9 @@ private actor RecordingDiagnosticGatewayMeasurer: DiagnosticGatewayMeasuring {
     }
 }
 
-private actor SequencedDiagnosticRouteSource: DiagnosticRouteSourcing {
-    private let values: [DiagnosticRouteSelection]
-    private var index = 0
+actor SequencedDiagnosticRouteSource: DiagnosticRouteSourcing {
+    let values: [DiagnosticRouteSelection]
+    var index = 0
     private(set) var invocationCount = 0
 
     init(values: [DiagnosticRouteSelection]) {
@@ -5103,7 +1537,7 @@ private actor SequencedDiagnosticRouteSource: DiagnosticRouteSourcing {
     }
 }
 
-private struct StubNetworkInterfaceSnapshotSource: NetworkInterfaceSnapshotSourcing {
+struct StubNetworkInterfaceSnapshotSource: NetworkInterfaceSnapshotSourcing {
     let interfaces: [NetworkInterfaceInfo]
 
     func capture(cycleID: UUID) async -> NetworkInterfaceSnapshot {
@@ -5115,7 +1549,7 @@ private struct StubNetworkInterfaceSnapshotSource: NetworkInterfaceSnapshotSourc
     }
 }
 
-private struct StubGatewayLatencyProvider: GatewayLatencyProviding {
+struct StubGatewayLatencyProvider: GatewayLatencyProviding {
     let result: GatewayLatencyResult
 
     func measure(routerIP: String?) async -> GatewayLatencyResult {
@@ -5123,7 +1557,7 @@ private struct StubGatewayLatencyProvider: GatewayLatencyProviding {
     }
 }
 
-private func makeNetworkInterface(name: String = "en0", router: String?) -> NetworkInterfaceInfo {
+func makeNetworkInterface(name: String = "en0", router: String?) -> NetworkInterfaceInfo {
     NetworkInterfaceInfo(
         interfaceName: name,
         hardwareMAC: "00:11:22:33:44:55",
@@ -5142,7 +1576,7 @@ private func makeNetworkInterface(name: String = "en0", router: String?) -> Netw
     )
 }
 
-private func makeDiagnosticContext(
+func makeDiagnosticContext(
     pathState: NetworkPathState?,
     route: DiagnosticRouteSelection,
     interfaces: [NetworkInterfaceInfo] = []
